Refactor ViewState into extensions for better maintainability
- Split ViewState.swift into 12 focused extension files
- ViewState+Types.swift - Supporting types and enums
- ViewState+Memory.swift - Memory management and cleanup
- ViewState+WebSocketEvents.swift - WebSocket event handling
- ViewState+UsersAndDms.swift - User and DM loading
- ViewState+Navigation.swift - Navigation and selection
- ViewState+Unreads.swift - Unread count management
- ViewState+Auth.swift - Authentication logic
- ViewState+ServerCache.swift - Server cache persistence
- ViewState+ReadyEvent.swift - Ready event processing
- ViewState+Notifications.swift - Push notifications
- ViewState+QueuedMessages.swift - Message queue handling
- ViewState+DMChannel.swift - DM channel operations

This improves code organization, maintainability, and makes it easier to navigate and modify specific functionality.
