+++ Revolt/ViewState.swift
@@ -12,271 +12,6 @@ import Darwin
 import Network
 
 
-enum UserStateError: Error {
-    case signInError
-    case signOutError
-}
-
-enum LoginState {
-    case Success
-    case Mfa(ticket: String, methods: [String])
-    case Disabled
-    case Invalid
-    case Onboarding
-}
-
-struct LoginSuccess: Decodable {
-    let result: String
-    let _id: String
-    let user_id: String
-    let token: String
-    let name: String
-}
-
-struct LoginMfa: Decodable {
-    let result: String
-    let ticket: String
-    let allowed_methods: [String]
-}
-
-struct LoginDisabled: Decodable {
-    let result: String
-    let user_id: String
-}
-
-enum LoginResponse {
-    case Success(LoginSuccess)
-    case Mfa(LoginMfa)
-    case Disabled(LoginDisabled)
-}
-
-extension LoginResponse: Decodable {
-    enum CodingKeys: String, CodingKey { case result }
-    enum Tag: String, Decodable { case Success, MFA, Disabled }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let singleValueContainer = try decoder.singleValueContainer()
-        
-        switch try container.decode(Tag.self, forKey: .result) {
-        case .Success:
-            self = .Success(try singleValueContainer.decode(LoginSuccess.self))
-        case .MFA:
-            self = .Mfa(try singleValueContainer.decode(LoginMfa.self))
-        case .Disabled:
-            self = .Disabled(try singleValueContainer.decode(LoginDisabled.self))
-        }
-    }
-}
-
-enum ConnectionState {
-    case connecting, connected, signedOut
-}
-
-
-enum MainSelection: Hashable, Codable {
-    case server(String)
-    case dms
-    case discover
-    
-    var id: String? {
-        switch self {
-        case .server(let id):
-            id
-        case .dms:
-            nil
-        case .discover:
-            nil
-        }
-    }
-}
-
-enum ChannelSelection: Hashable, Codable {
-    case channel(String)
-    case home
-    case friends
-    case noChannel
-    
-    var id: String? {
-        switch self {
-        case .channel(let id): id
-        default: nil
-        }
-    }
-}
-
-enum NavigationDestination: Hashable, Codable {
-    case discover
-    case settings
-    case about_settings
-    case developer_settings
-    case server_settings(String)
-    case channel_info(String,String?)
-    case add_members_to_channel(String)
-    case channel_settings(String)
-    case add_friend
-    case create_group([String])
-    case channel_search(String)
-    case invite(String)
-    case maybeChannelView
-    case create_group_name
-    case create_group_add_memebers(String)
-    case report(Types.User?,String?,String?)
-    case channel_overview_setting(String,String?)
-    case server_channel_overview_setting(String,String)
-    case server_role_setting(String)
-    case server_overview_settings(String)
-    case server_channels(String)
-    case server_category(String,String)
-    case channel_category_create(String, ChannelCategoryCreateType)
-    case profile_setting
-    case server_emoji_settings(String)
-    case validate_password_view(ValidatePasswordReason)
-    case show_recovery_codes(String, Bool)
-    case enable_authenticator_app(String)
-    case blocked_users_view
-    case user_settings
-    case role_setting(serverId:String, channelId: String, roleId : String, roleTitle : String, value:ChannelRolePermissionsSettings.Value)
-    case server_members_view(String)
-    case member_permissions(String, Member)
-    case server_invites(String)
-    case server_banned_users(String)
-    case create_server_role(serverId : String)
-    case default_role_settings(serverId : String)
-    case role_settings(serverId : String, roleId : String)
-    case channel_permissions_settings(serverId : String?, channelId : String)
-    case sessions_settings
-    case username_view
-    case change_email_view
-    case change_password_view
-}
-
-struct UserMaybeMember: Identifiable {
-    var user: Types.User
-    var member: Member?
-    
-    var id: String { user.id }
-    
-    var avatar: File? {
-        return member?.avatar ?? user.avatar
-    }
-    
-    var nickname: String? {
-        return member?.nickname ?? user.display_name ?? user.username
-    }
-}
-
-class QueuedMessage: ObservableObject {
-    let nonce: String
-    let replies: [Revolt.ApiReply]
-    let content: String
-    let author: String // User ID of the sender
-    let channel: String // Channel ID
-    let timestamp: Date // When the message was sent locally
-    let hasAttachments: Bool // Whether this message has attachments (affects when to show optimistically)
-    let attachmentData: [(Data, String)] // Original attachment data for progress tracking
-    @Published var uploadProgress: [String: Double] = [:] // Progress per attachment (filename -> progress 0.0-1.0)
-    @Published var isUploading: Bool = false // Whether currently uploading
-    
-    init(nonce: String, replies: [Revolt.ApiReply], content: String, author: String, channel: String, timestamp: Date, hasAttachments: Bool, attachmentData: [(Data, String)] = []) {
-        self.nonce = nonce
-        self.replies = replies
-        self.content = content
-        self.author = author
-        self.channel = channel
-        self.timestamp = timestamp
-        self.hasAttachments = hasAttachments
-        self.attachmentData = attachmentData
-        self.uploadProgress = [:]
-        self.isUploading = hasAttachments
-        
-        // Initialize progress for each attachment
-        for (_, filename) in attachmentData {
-            self.uploadProgress[filename] = 0.0
-        }
-    }
-    
-    // Helper function to determine content type from filename
-    private func getContentType(for filename: String) -> String {
-        let lowercaseFilename = filename.lowercased()
-        
-        // Image types
-        if lowercaseFilename.hasSuffix(".jpg") || lowercaseFilename.hasSuffix(".jpeg") {
-            return "image/jpeg"
-        } else if lowercaseFilename.hasSuffix(".png") {
-            return "image/png"
-        } else if lowercaseFilename.hasSuffix(".gif") {
-            return "image/gif"
-        } else if lowercaseFilename.hasSuffix(".webp") {
-            return "image/webp"
-        } else if lowercaseFilename.hasSuffix(".bmp") {
-            return "image/bmp"
-        } else if lowercaseFilename.hasSuffix(".svg") {
-            return "image/svg+xml"
-        }
-        // Video types
-        else if lowercaseFilename.hasSuffix(".mp4") {
-            return "video/mp4"
-        } else if lowercaseFilename.hasSuffix(".mov") {
-            return "video/quicktime"
-        } else if lowercaseFilename.hasSuffix(".avi") {
-            return "video/x-msvideo"
-        } else if lowercaseFilename.hasSuffix(".webm") {
-            return "video/webm"
-        }
-        // Audio types
-        else if lowercaseFilename.hasSuffix(".mp3") {
-            return "audio/mpeg"
-        } else if lowercaseFilename.hasSuffix(".wav") {
-            return "audio/wav"
-        } else if lowercaseFilename.hasSuffix(".m4a") {
-            return "audio/mp4"
-        }
-        // Default
-        else {
-            return "application/octet-stream"
-        }
-    }
-    
-    // Convert to a temporary Message object for display
-    func toTemporaryMessage() -> Message {
-        // Create temporary File objects for display if uploading
-        let tempAttachments: [Types.File]? = hasAttachments && isUploading ? 
-            attachmentData.map { (data, filename) in
-                // Determine content type based on file extension
-                let contentType = getContentType(for: filename)
-                
-                return Types.File(
-                    id: "\(nonce)_\(filename)", // Temporary ID
-                    tag: "attachments", // Required tag parameter
-                    size: Int64(data.count),
-                    filename: filename,
-                    metadata: Types.FileMetadata.file(Types.SimpleMetadata()), // Correct metadata type
-                    content_type: contentType
-                )
-            } : nil
-        
-        return Message(
-            id: nonce, // Use nonce as temporary ID
-            content: content,
-            author: author,
-            channel: channel,
-            system: nil,
-            attachments: tempAttachments,
-            mentions: nil,
-            replies: replies.isEmpty ? nil : replies.map { $0.id },
-            edited: nil,
-            masquerade: nil,
-            interactions: nil,
-            reactions: nil,
-            user: nil,
-            member: nil,
-            embeds: nil,
-            webhook: nil
-        )
-    }
-}
-
 @MainActor
 public class ViewState: ObservableObject {
     var cancellables = Set<AnyCancellable>()
@@ -444,11 +179,11 @@ public class ViewState: ObservableObject {
     @Published var allDmChannelIds: [String] = [] // All DM channel IDs in order
     @Published var loadedDmBatches: Set<Int> = [] // Track which batches are loaded
     let dmBatchSize = 15 // Load 15 DMs per batch
-    private var isLoadingDmBatch = false
-    private var isDmListInitialized = false // Track if DM list has been initialized
+    internal var isLoadingDmBatch = false
+    internal var isDmListInitialized = false // Track if DM list has been initialized
     
     // SIMPLE LAZY LOADING: Just load more as needed, no complex virtual scrolling
-    private let maxLoadedBatches = 50 // Maximum 50 batches (750 DMs) with increased memory limits
+    internal let maxLoadedBatches = 50 // Maximum 50 batches (750 DMs) with increased memory limits
     @Published var emojis: [String: Emoji] {
         didSet {
             // OPTIMIZED: Move encoding to background thread with debouncing
@@ -517,10 +252,10 @@ public class ViewState: ObservableObject {
     }
     
     // MEMORY MANAGEMENT: Add debouncing for UserDefaults saves
-    private var saveWorkItems: [String: DispatchWorkItem] = [:]
+    internal var saveWorkItems: [String: DispatchWorkItem] = [:]
     private let saveDebounceInterval: TimeInterval = 2.0 // Save after 2 seconds of no changes
     private let cleanupTriggeredAt = 800 // Start cleanup when 80% full (legacy, not used)
-    private let maxChannelsInMemory = 2000 // Maximum channels to keep in memory (increased to load all servers)
+    internal let maxChannelsInMemory = 2000 // Maximum channels to keep in memory (increased to load all servers)
     
     @Published var unreadsVersion: UUID = UUID()
     @Published var currentUserSheet: UserMaybeMember? = nil
@@ -610,7 +345,7 @@ public class ViewState: ObservableObject {
     @Published var lastInviteServerContext: String? = nil
     
     // MEMORY MANAGEMENT: Track previous channel for cleanup
-    private var previousChannelId: String? = nil
+    internal var previousChannelId: String? = nil
     
     // Loading state for channels when messages are being fetched
     @Published var isLoadingChannelMessages: Bool = false
@@ -621,14 +356,14 @@ public class ViewState: ObservableObject {
     var userSettingsStore: UserSettingsData
     
     // MEMORY MANAGEMENT: Configuration for aggressive cleanup
-    private let maxMessagesInMemory = 7000 // Maximum messages to keep in memory
-    private let maxUsersInMemory = 2000 // Increased to handle 300+ messages with users
-    private let maxChannelMessages = 800 // Maximum messages per channel (reduced for better memory management)
+    internal let maxMessagesInMemory = 7000 // Maximum messages to keep in memory
+    internal let maxUsersInMemory = 2000 // Increased to handle 300+ messages with users
+    internal let maxChannelMessages = 800 // Maximum messages per channel (reduced for better memory management)
     private let maxServersInMemory = 50 // Maximum servers to keep in memory
     
     // PRELOADING CONTROL: Configuration for automatic message preloading
     /// Set to false to disable all automatic message preloading when entering servers/channels
-    private let enableAutomaticPreloading = false // DISABLED: No automatic preloading
+    internal let enableAutomaticPreloading = false // DISABLED: No automatic preloading
     
     // MEMORY MANAGEMENT: Helper methods
     /// Debounced save helper that defers heavy JSON encoding to the background queue
@@ -667,278 +402,6 @@ public class ViewState: ObservableObject {
         }
     }
     
-    @MainActor
-    private func enforceMemoryLimits() {
-        // CRITICAL FIX: Completely disable enforceMemoryLimits to prevent infinite loop and black messages
-        // print("🚫 MEMORY_CLEANUP: enforceMemoryLimits DISABLED to prevent infinite loop and black messages")
-        return
-        
-        // Check current memory usage
-        let currentMemoryMB = getCurrentMemoryUsage()
-        
-        // EMERGENCY MEMORY RESET if over 4GB
-        if currentMemoryMB > 4000 {
-            // print("🚨 EMERGENCY: Memory over 4GB (\(currentMemoryMB)MB)! Performing complete reset!")
-            
-            // Clear everything except current user
-            let currentUserId = currentUser?.id
-            let currentUserObject = currentUser
-            
-            messages.removeAll()
-            channelMessages.removeAll()
-            users.removeAll()
-            channels.removeAll()
-            servers.removeAll()
-            members.removeAll()
-            dms.removeAll()
-            emojis.removeAll()
-            unreads.removeAll()
-            
-            // Restore only current user
-            if let currentUserId = currentUserId, let currentUserObject = currentUserObject {
-                users[currentUserId] = currentUserObject
-                currentUser = currentUserObject
-            }
-            
-            // print("🚨 EMERGENCY RESET COMPLETED! Memory should now be minimal.")
-            return
-        }
-        
-                    // AGGRESSIVE MEMORY CLEANUP if over 2GB (increased threshold for better performance)
-            if currentMemoryMB > 2000 {
-                // print("🚨 AGGRESSIVE CLEANUP: Memory over 2GB (\(currentMemoryMB)MB)!")
-                
-                // VIRTUAL SCROLLING PROTECTION: Skip aggressive cleanup if in DM view
-                if currentSelection == .dms {
-                    // print("🔄 VIRTUAL_DM: Skipping aggressive cleanup - user is in DM view with Virtual Scrolling active")
-                    return
-                }
-                
-                                    // CRITICAL FIX: Keep ALL users to prevent black messages - only clear non-essential data
-                    // print("🚨 EMERGENCY: Keeping ALL users to prevent black messages")
-                    // Don't touch users at all - they are needed for message display
-                
-                // Keep only last 100 messages
-                let sortedMessages = messages.sorted { $0.value.id > $1.value.id }
-                let recentMessages = Array(sortedMessages.prefix(100))
-                messages = Dictionary(uniqueKeysWithValues: recentMessages)
-                
-                // Keep ALL DMs and current channel
-                var channelsToKeep: [String: Channel] = [:]
-                
-                if case .channel(let currentChannelId) = currentChannel {
-                    if let currentChannel = channels[currentChannelId] {
-                        channelsToKeep[currentChannelId] = currentChannel
-                    }
-                }
-                
-                // Keep ALL DM and Group DM channels (don't limit to 10)
-                for channel in channels.values {
-                    switch channel {
-                    case .dm_channel:
-                        channelsToKeep[channel.id] = channel
-                    case .group_dm_channel:
-                        channelsToKeep[channel.id] = channel
-                    default:
-                        break
-                    }
-                }
-                
-                channels = channelsToKeep
-                
-                // Clear all but essential channel messages
-                for (channelId, _) in channelMessages {
-                    if channelsToKeep[channelId] != nil {
-                        channelMessages[channelId] = Array((channelMessages[channelId] ?? []).suffix(10))
-                    } else {
-                        channelMessages.removeValue(forKey: channelId)
-                    }
-                }
-                
-                // Keep more servers in emergency cleanup (increased from 5 to 20)
-                let topServers = Array(servers.prefix(20))
-                servers = OrderedDictionary(uniqueKeysWithValues: topServers)
-                
-                // FIX: Don't clear DM list state during aggressive cleanup
-                if isDmListInitialized {
-                    // Keep DM list state intact, just reinitialize it
-                    reinitializeDmListFromCache()
-                }
-                
-                // print("🚨 AGGRESSIVE CLEANUP COMPLETED!")
-                return
-            }
-        
-        // NORMAL CLEANUP: Remove excess messages
-        if messages.count > maxMessagesInMemory {
-            // print("🧠 MEMORY: Enforcing message limit. Current: \(messages.count), Max: \(maxMessagesInMemory)")
-            
-            // Get all message IDs sorted by timestamp (older first)
-            let sortedMessageIds = messages.keys.sorted { id1, id2 in
-                let date1 = createdAt(id: id1)
-                let date2 = createdAt(id: id2)
-                return date1 < date2
-            }
-            
-            // Calculate how many messages to remove
-            let messagesToRemove = messages.count - maxMessagesInMemory
-            let idsToRemove = Array(sortedMessageIds.prefix(messagesToRemove))
-            
-            // Remove messages
-            for id in idsToRemove {
-                messages.removeValue(forKey: id)
-            }
-            
-            // Clean up channel message references
-            for (channelId, messageIds) in channelMessages {
-                let filteredIds = messageIds.filter { !idsToRemove.contains($0) }
-                if filteredIds.count != messageIds.count {
-                    channelMessages[channelId] = filteredIds
-                }
-            }
-            
-            // print("🧠 MEMORY: Removed \(messagesToRemove) old messages")
-        }
-        
-        // AGGRESSIVE CHANNEL MESSAGE CLEANUP
-        for (channelId, messageIds) in channelMessages {
-            if messageIds.count > maxChannelMessages {
-                let trimmedIds = Array(messageIds.suffix(maxChannelMessages))
-                channelMessages[channelId] = trimmedIds
-                // print("🧠 MEMORY: Trimmed channel \(channelId) messages from \(messageIds.count) to \(trimmedIds.count)")
-            }
-        }
-    }
-    
-    // DISABLED: Smart message cleanup based on current channel and loading direction
-    @MainActor
-    private func smartMessageCleanup() {
-        // CRITICAL FIX: Disable message cleanup to prevent black messages
-        // print("🚫 MEMORY_CLEANUP: smartMessageCleanup DISABLED to prevent black messages")
-        return
-    }
-    
-    // Smart user cleanup to prevent excessive memory usage - DISABLED to prevent black messages
-    @MainActor
-    private func smartUserCleanup() {
-        // CRITICAL FIX: Completely disable user cleanup to prevent black messages
-        // print("🧠 MEMORY: User cleanup DISABLED to prevent black messages. Current users: \(users.count)")
-        
-        // Only log warning if we have too many users, but don't clean them up
-        if users.count > maxUsersInMemory {
-            // print("⚠️ MEMORY WARNING: \(users.count) users exceed limit of \(maxUsersInMemory), but cleanup is disabled")
-        }
-        
-        return // Exit early, no cleanup
-    }
-    
-    @MainActor
-    private func cleanupMemory() {
-        // CRITICAL FIX: Completely disable cleanupMemory to prevent infinite loop and black messages
-        // print("🚫 MEMORY: cleanupMemory DISABLED to prevent infinite loop and black messages")
-        return
-    }
-    
-    // Smart channel cleanup to prevent excessive memory usage
-    @MainActor
-    private func smartChannelCleanup() {
-        // FIX: Don't cleanup when in DM view or when DM list is being displayed
-        if currentSelection == .dms {
-            // print("🧠 MEMORY: Skipping channel cleanup - in DM view")
-            return
-        }
-        
-        // Clean up channels that haven't been accessed
-        if channels.count > maxChannelsInMemory {
-            // print("🧠 MEMORY: Enforcing channel limit. Current: \(channels.count), Max: \(maxChannelsInMemory)")
-            
-            var essentialChannelIds = Set<String>()
-            
-            // Keep current channel
-            if case .channel(let channelId) = currentChannel {
-                essentialChannelIds.insert(channelId)
-            }
-            
-            // ALWAYS keep ALL DM and Group DM channels (NEVER remove these!)
-            for channel in channels.values {
-                switch channel {
-                case .dm_channel:
-                    essentialChannelIds.insert(channel.id)
-                case .group_dm_channel:
-                    essentialChannelIds.insert(channel.id)
-                default:
-                    break
-                }
-            }
-            
-            // Keep channels in current server only
-            if case .server(let serverId) = currentSelection {
-                if let server = servers[serverId] {
-                    // Only keep current server's channels to free memory
-                    for channelId in server.channels {
-                        if essentialChannelIds.count < maxChannelsInMemory {
-                            essentialChannelIds.insert(channelId)
-                        }
-                    }
-                }
-            }
-            
-            // Remove non-essential channels (only server channels from other servers)
-            let channelsToRemove = channels.keys.filter { channelId in
-                if essentialChannelIds.contains(channelId) {
-                    return false
-                }
-                
-                // Check if it's a server text channel from non-current server
-                if let channel = channels[channelId] {
-                    switch channel {
-                    case .text_channel:
-                        return true // Remove server channels that aren't essential
-                    case .voice_channel:
-                        return true // Remove voice channels
-                    default:
-                        return false // Never remove DMs
-                    }
-                }
-                return false
-            }
-            
-            for channelId in channelsToRemove {
-                channels.removeValue(forKey: channelId)
-                channelMessages.removeValue(forKey: channelId)
-            }
-            
-            // print("🧠 MEMORY: Removed \(channelsToRemove.count) non-essential channels (kept all DMs)")
-        }
-        
-        // Clean up empty channel message arrays for server channels only
-        let emptyChannels = channelMessages.filter { channelId, messages in
-            if messages.isEmpty {
-                // Check if it's a DM - if so, don't remove
-                if let channel = channels[channelId] {
-                    switch channel {
-                    case .dm_channel, .group_dm_channel:
-                        return false // Never remove DM message arrays
-                    default:
-                        return true // Can remove server channel message arrays if empty
-                    }
-                }
-                return true
-            }
-            return false
-        }.map { $0.key }
-        
-        if emptyChannels.count > 100 {
-            // print("🧠 MEMORY: Cleaning up \(emptyChannels.count) empty server channel message arrays")
-            for channelId in emptyChannels {
-                // Only remove if it's not current channel and not a DM
-                if case .channel(let currentChannelId) = currentChannel, currentChannelId == channelId {
-                    continue
-                }
-                channelMessages.removeValue(forKey: channelId)
-            }
-        }
-    }
     
     // Clear messages when leaving a channel to free memory, keeping only last 100 messages
     @MainActor
@@ -979,22 +442,6 @@ public class ViewState: ObservableObject {
         // print("🧠 MEMORY: Channel cleanup completed for: \(channelId)")
     }
     
-    // Helper function to clean up messages that don't belong to any channel
-    @MainActor
-    private func cleanupOrphanedMessages() {
-        let allChannelMessageIds = Set(channelMessages.values.flatMap { $0 })
-        let messagesToRemove = messages.keys.filter { messageId in
-            !allChannelMessageIds.contains(messageId)
-        }
-        
-        for messageId in messagesToRemove {
-            messages.removeValue(forKey: messageId)
-        }
-        
-        if !messagesToRemove.isEmpty {
-            // print("🧠 MEMORY: Cleaned up \(messagesToRemove.count) orphaned messages")
-        }
-    }
     
     // Clean all channel messages to keep only last 100 per channel
     @MainActor
@@ -1034,76 +481,6 @@ public class ViewState: ObservableObject {
     func setChannelLoadingState(isLoading: Bool) {
         isLoadingChannelMessages = isLoading
         // print("🧠 LOADING: Channel loading state changed to: \(isLoading)")
-    }
-    
-    // Handle channel change and memory cleanup  
-    @MainActor
-    private func handleChannelChange(from previousChannelId: String?, to newChannel: ChannelSelection) {
-        // Extract previous channel ID from current channel before change
-        let actualPreviousChannelId: String?
-        if case .channel(let channelId) = currentChannel {
-            actualPreviousChannelId = channelId
-        } else {
-            actualPreviousChannelId = nil
-        }
-        
-        // Extract new channel ID
-        let newChannelId: String?
-        if case .channel(let channelId) = newChannel {
-            newChannelId = channelId
-        } else {
-            newChannelId = nil
-        }
-        
-        // Clear messages from previous channel when switching channels
-        if let actualPreviousChannelId = actualPreviousChannelId, actualPreviousChannelId != newChannelId {
-            // print("🧠 MEMORY: Switching channels from \(actualPreviousChannelId) to \(newChannelId ?? "none")")
-            clearChannelMessages(channelId: actualPreviousChannelId)
-            
-            // CRITICAL: Clear target message ID when switching channels to prevent re-targeting
-            print("🎯 CHANNEL_CHANGE: Clearing currentTargetMessageId when switching channels")
-            currentTargetMessageId = nil
-        }
-        
-        // Update previous channel ID for next time
-        self.previousChannelId = newChannelId
-    }
-    
-    // Handle path changes to detect when leaving channel view
-    @MainActor
-    private func handlePathChange(oldPath: [NavigationDestination], newPath: [NavigationDestination]) {
-        let wasInChannelView = oldPath.contains { destination in
-            if case .maybeChannelView = destination {
-                return true
-            }
-            return false
-        }
-        
-        let isInChannelView = newPath.contains { destination in
-            if case .maybeChannelView = destination {
-                return true
-            }
-            return false
-        }
-        
-        // Clear messages when leaving channel view to free memory
-        if wasInChannelView && !isInChannelView {
-            if case .channel(let channelId) = currentChannel {
-                // print("🧠 MEMORY: Left channel view, clearing messages for channel: \(channelId) to free memory")
-                clearChannelMessages(channelId: channelId)
-            }
-        }
-        
-        // If we're entering channel view, check if we need to show loading
-        if !wasInChannelView && isInChannelView {
-            if case .channel(let channelId) = currentChannel {
-                let hasMessages = (channelMessages[channelId]?.count ?? 0) > 0
-                if !hasMessages {
-                    // print("🧠 LOADING: Entering channel with no messages, showing loading state")
-                    setChannelLoadingState(isLoading: true)
-                }
-            }
-        }
     }
     
     static func decodeUserDefaults<T: Decodable>(forKey key: String, withDecoder decoder: JSONDecoder) throws -> T? {
@@ -1271,71 +648,6 @@ public class ViewState: ObservableObject {
     /// Set to track which channels have been preloaded to avoid duplicates
     var preloadedChannels: Set<String> = Set()
     
-    /// Preloads messages for important channels when the app starts or WebSocket reconnects
-    @MainActor
-    private func preloadImportantChannels() async {
-        // Wait a bit for the WebSocket to fully authenticate
-        try? await Task.sleep(nanoseconds: 2_000_000_000) // 2 seconds
-        
-        // Only preload if user is authenticated and WebSocket is connected
-        guard sessionToken != nil, currentUser != nil, state == .connected else {
-            print("🚀 PRELOAD: Skipping preload - user not authenticated or not connected (state: \(state))")
-            return
-        }
-        
-        // SMART PRELOADING: Get channels from user's current server and DMs
-        var channelsToPreload: [String] = []
-        
-        // Add current server's channels
-        if case .server(let serverId) = currentSelection,
-           let server = servers[serverId] {
-            // Add first few text channels from current server
-            let textChannels = server.channels.compactMap { channelId in
-                if case .text_channel(_) = channels[channelId] {
-                    return channelId
-                }
-                return nil
-            }.prefix(3) // Preload first 3 text channels
-            
-            channelsToPreload.append(contentsOf: textChannels)
-            print("🚀 PRELOAD: Added \(textChannels.count) channels from current server \(serverId)")
-        }
-        
-        // Add active DM channels
-        let activeDMs = dms.compactMap { channel -> String? in
-            switch channel {
-            case .dm_channel(let dm):
-                return dm.active ? dm.id : nil
-            case .group_dm_channel(let group):
-                return group.id
-            default:
-                return nil
-            }
-        }.prefix(5) // Preload first 5 DMs
-        
-        channelsToPreload.append(contentsOf: activeDMs)
-        print("🚀 PRELOAD: Added \(activeDMs.count) DM channels")
-        
-        // Always include the specific channel mentioned by user
-        let specificChannelId = "01J7QTT66242A7Q26A2FH5TD48"
-        if !channelsToPreload.contains(specificChannelId) {
-            channelsToPreload.append(specificChannelId)
-            print("🚀 PRELOAD: Added specific channel \(specificChannelId)")
-        }
-        
-        print("🚀 PRELOAD: Starting preload for \(channelsToPreload.count) channels")
-        
-        // Preload channels in parallel for better performance
-        await withTaskGroup(of: Void.self) { group in
-            for channelId in channelsToPreload {
-                group.addTask {
-                    await self.preloadChannel(channelId: channelId)
-        }
-            }
-        }
-        
-        print("🚀 PRELOAD: Completed preloading \(channelsToPreload.count) channels")
-    }
     
     /// Public method to manually trigger preload for important channels
     @MainActor
@@ -1343,11 +655,6 @@ public class ViewState: ObservableObject {
         await preloadImportantChannels()
     }
     
-    /// Public method to preload a specific channel by ID
-    @MainActor
-    public func preloadSpecificChannel(channelId: String) async {
-        await preloadChannel(channelId: channelId)
-    }
     
     /// Public method to preload multiple channels by IDs
     @MainActor
@@ -1359,7 +666,7 @@ public class ViewState: ObservableObject {
     
     /// Preloads messages for a specific channel
     @MainActor
-    private func preloadChannel(channelId: String) async {
+    internal func preloadChannel(channelId: String) async {
         // Check if channel has already been preloaded
         if preloadedChannels.contains(channelId) {
             print("🚀 PRELOAD: Channel \(channelId) has already been preloaded, skipping")
@@ -1439,27 +746,12 @@ public class ViewState: ObservableObject {
     }
     
     // MEMORY MANAGEMENT: Periodic cleanup timer
-    private var memoryCleanupTimer: Timer?
-    private var memoryMonitorTimer: Timer?
-    
-    private func startPeriodicMemoryCleanup() {
-        memoryCleanupTimer?.invalidate()
-        
-        // Clean up memory every 15 seconds for better stability
-        memoryCleanupTimer = Timer.scheduledTimer(withTimeInterval: 15.0, repeats: true) { [weak self] _ in
-            guard let self = self else { return }
-            
-            Task { @MainActor [weak self] in
-                self?.cleanupMemory()
-            }
-        }
-        
-        // Start memory monitoring
-        startMemoryMonitoring()
-    }
+    internal var memoryCleanupTimer: Timer?
+    internal var memoryMonitorTimer: Timer?
+
     
     // MEMORY MONITORING: Track memory usage
-    private func startMemoryMonitoring() {
+    internal func startMemoryMonitoring() {
         memoryMonitorTimer?.invalidate()
         
         memoryMonitorTimer = Timer.scheduledTimer(withTimeInterval: 5.0, repeats: true) { [weak self] _ in
@@ -1503,7 +795,7 @@ public class ViewState: ObservableObject {
         }
     }
     
-    private func getCurrentMemoryUsage() -> Double {
+    func getCurrentMemoryUsage() -> Double {
         var info = mach_task_basic_info()
         var count = mach_msg_type_number_t(MemoryLayout<mach_task_basic_info>.size) / 4
         
@@ -1560,56 +852,10 @@ public class ViewState: ObservableObject {
         return Date(timeIntervalSince1970: TimeInterval(timestamp) / 1000.0)
     }
     
-    /// INSTANT cleanup for a specific channel when leaving it - NO DELAYS
-    @MainActor
-    func cleanupChannelFromMemory(channelId: String, preserveForNavigation: Bool = false) {
-        let startTime = CFAbsoluteTimeGetCurrent()
-        print("⚡ VIEWSTATE_INSTANT_CLEANUP: Starting IMMEDIATE cleanup for channel \(channelId)")
-        
-        let initialMessageCount = messages.count
-        let initialUserCount = users.count
-        let channelMessageCount = channelMessages[channelId]?.count ?? 0
-        
-        // 1. IMMEDIATE: Clear channel messages list
-        channelMessages.removeValue(forKey: channelId)
-        
-        // 2. IMMEDIATE: Remove all message objects for this channel
-        let messagesToRemove = messages.keys.filter { messageId in
-            if let message = messages[messageId] {
-                return message.channel == channelId
-            }
-            return false
-        }
-        
-        for messageId in messagesToRemove {
-            messages.removeValue(forKey: messageId)
-        }
-        
-        // 3. IMMEDIATE: Clear all related data
-        currentlyTyping.removeValue(forKey: channelId)
-        preloadedChannels.remove(channelId)
-        atTopOfChannel.remove(channelId)
-        
-        // 4. IMMEDIATE: Clean up users if not preserving for navigation
-        if !preserveForNavigation {
-            cleanupUnusedUsersInstant(excludingChannelId: channelId)
-        }
-        
-        // 5. IMMEDIATE: Force garbage collection
-        _ = messages.count + users.count + channelMessages.count
-        
-        let finalMessageCount = messages.count
-        let finalUserCount = users.count
-        let endTime = CFAbsoluteTimeGetCurrent()
-        let duration = (endTime - startTime) * 1000
-        
-        print("⚡ VIEWSTATE_INSTANT_CLEANUP: Completed in \(String(format: "%.2f", duration))ms")
-        print("⚡ FREED: \(initialMessageCount - finalMessageCount) messages, \(initialUserCount - finalUserCount) users, \(channelMessageCount) channel messages")
-    }
     
     /// INSTANT cleanup of unused users - NO DELAYS
     @MainActor
-    private func cleanupUnusedUsersInstant(excludingChannelId: String) {
+    internal func cleanupUnusedUsersInstant(excludingChannelId: String) {
         let startTime = CFAbsoluteTimeGetCurrent()
         
         var usersToKeep = Set<String>()
@@ -1674,133 +920,6 @@ public class ViewState: ObservableObject {
         print("⚡ USER_INSTANT_CLEANUP: Removed \(usersToRemove.count) users in \(String(format: "%.2f", duration))ms (\(initialUserCount) -> \(finalUserCount))")
     }
     
-    /// Clean up users that are no longer needed after leaving a channel
-    @MainActor
-    private func cleanupUnusedUsers(excludingChannelId: String) {
-        print("👥 USER_CLEANUP: Starting cleanup of unused users")
-        
-        var usersToKeep = Set<String>()
-        
-        // Always keep current user
-        if let currentUserId = currentUser?.id {
-            usersToKeep.insert(currentUserId)
-        }
-        
-        // Keep users from all active channels (except the one we're leaving)
-        for (otherChannelId, messageIds) in channelMessages {
-            if otherChannelId == excludingChannelId { continue }
-            
-            // Keep users from channel recipients (for DMs)
-            if let channel = channels[otherChannelId] {
-                usersToKeep.formUnion(channel.recipients)
-            }
-            
-            // Keep message authors and mentioned users
-            for messageId in messageIds {
-                if let message = messages[messageId] {
-                    usersToKeep.insert(message.author)
-                    if let mentions = message.mentions {
-                        usersToKeep.formUnion(mentions)
-                    }
-                }
-            }
-        }
-        
-        // Keep users from servers (owners and members)
-        for server in servers.values {
-            usersToKeep.insert(server.owner)
-            if let serverMembers = members[server.id] {
-                usersToKeep.formUnion(serverMembers.keys)
-            }
-        }
-        
-        // Keep users from DM list
-        for dm in dms {
-            usersToKeep.formUnion(dm.recipients)
-        }
-        
-        // Remove users that are no longer needed
-        let initialUserCount = users.count
-        let usersToRemove = users.keys.filter { userId in
-            !usersToKeep.contains(userId) && userId != currentUser?.id
-        }
-        
-        for userId in usersToRemove {
-            users.removeValue(forKey: userId)
-            
-            // Also remove from members if they exist
-            for serverId in members.keys {
-                members[serverId]?.removeValue(forKey: userId)
-            }
-        }
-        
-        let finalUserCount = users.count
-        print("👥 USER_CLEANUP: Removed \(usersToRemove.count) unused users (\(initialUserCount) -> \(finalUserCount))")
-    }
-    
-    /// INSTANT force memory cleanup - IMMEDIATE execution
-    @MainActor
-    func forceMemoryCleanup() {
-        let startTime = CFAbsoluteTimeGetCurrent()
-        print("⚡ FORCE_INSTANT_CLEANUP: Starting IMMEDIATE memory cleanup")
-        
-        let initialStats = (
-            messages: messages.count,
-            users: users.count,
-            channels: channelMessages.count
-        )
-        
-        // 1. IMMEDIATE: Enforce message limits aggressively
-        if messages.count > maxMessagesInMemory {
-            let sortedMessageIds = messages.keys.sorted { id1, id2 in
-                let date1 = createdAt(id: id1)
-                let date2 = createdAt(id: id2)
-                return date1 < date2
-            }
-            
-            let messagesToRemove = messages.count - maxMessagesInMemory
-            let idsToRemove = Array(sortedMessageIds.prefix(messagesToRemove))
-            
-            for id in idsToRemove {
-                messages.removeValue(forKey: id)
-            }
-            
-            // Clean up channel message references
-            for (channelId, messageIds) in channelMessages {
-                let filteredIds = messageIds.filter { !idsToRemove.contains($0) }
-                channelMessages[channelId] = filteredIds
-            }
-        }
-        
-        // 2. IMMEDIATE: Enforce user limits
-        if users.count > maxUsersInMemory {
-            cleanupUnusedUsersInstant(excludingChannelId: "")
-        }
-        
-        // 3. IMMEDIATE: Clean up empty channel message arrays
-        for (channelId, messageIds) in channelMessages {
-            if messageIds.isEmpty {
-                channelMessages.removeValue(forKey: channelId)
-            }
-        }
-        
-        // 4. IMMEDIATE: Force garbage collection
-        _ = messages.count + users.count + channelMessages.count
-        
-        let finalStats = (
-            messages: messages.count,
-            users: users.count,
-            channels: channelMessages.count
-        )
-        
-        let endTime = CFAbsoluteTimeGetCurrent()
-        let duration = (endTime - startTime) * 1000
-        
-        print("⚡ FORCE_INSTANT_CLEANUP: Completed in \(String(format: "%.2f", duration))ms")
-        print("   Messages: \(initialStats.messages) -> \(finalStats.messages)")
-        print("   Users: \(initialStats.users) -> \(finalStats.users)")
-        print("   Channels: \(initialStats.channels) -> \(finalStats.channels)")
-    }
     
     func setBaseUrlToHttp() {
         http.baseURL = baseURL ?? defaultBaseURL
@@ -1917,152 +1036,6 @@ public class ViewState: ObservableObject {
         return true
     }
     
-    func signIn(mfa_ticket: String, mfa_response: [String: String], callback: @escaping((LoginState) -> ())) async {
-        let body = ["mfa_ticket": mfa_ticket, "mfa_response": mfa_response, "friendly_name": "Revolt iOS"] as [String : Any]
-        
-        await innerSignIn(body, callback)
-    }
-    
-    func signIn(email: String, password: String, callback: @escaping((LoginState) -> ())) async {
-        let body = ["email": email, "password":  password, "friendly_name": "Revolt IOS"]
-        
-        await innerSignIn(body, callback)
-    }
-    
-    private func innerSignIn(_ body: [String: Any], _ callback: @escaping((LoginState) -> ())) async {
-        AF.request("\(http.baseURL)/auth/session/login", method: .post, parameters: body, encoding: JSONEncoding.default)
-            .responseData { response in
-                
-                switch response.result {
-                case .success(let data):
-                    if [401, 403, 500].contains(response.response!.statusCode) {
-                        return callback(.Invalid)
-                    }
-                    if let result = try? JSONDecoder().decode(LoginResponse.self, from: data){
-                        switch result {
-                        case .Success(let success):
-                            Task {
-                                self.isOnboarding = true
-                                self.currentSessionId = success._id
-                                self.sessionToken = success.token
-                                self.http.token = success.token
-                                
-                                await self.promptForNotifications()
-                                
-                                // If we already have a device notification token, try to upload it
-                                if let existingToken = self.deviceNotificationToken {
-                                    // print("📱 LOGIN_SUCCESS: Found existing device token, uploading...")
-                                    Task {
-                                        let response = await self.http.uploadNotificationToken(token: existingToken)
-                                        switch response {
-                                            case .success:
-                                                print("✅ LOGIN_SUCCESS: Successfully uploaded existing token")
-                                            case .failure(let error):
-                                                print("❌ LOGIN_SUCCESS: Failed to upload existing token: \(error)")
-                                                self.storePendingNotificationToken(existingToken)
-                                        }
-                                    }
-                                }
-                                
-                                do {
-                                    let onboardingState = try await self.http.checkOnboarding().get()
-                                    if onboardingState.onboarding {
-                                        self.isOnboarding = true
-                                        callback(.Onboarding)
-                                    } else {
-                                        self.isOnboarding = false
-                                        callback(.Success)
-                                        self.state = .connecting
-                                    }
-                                } catch {
-                                    self.isOnboarding = false
-                                    self.state = .connecting
-                                    return callback(.Success) // if the onboard check dies, just try to go for it
-                                }
-                            }
-                            
-                        case .Mfa(let mfa):
-                            return callback(.Mfa(ticket: mfa.ticket, methods: mfa.allowed_methods))
-                            
-                        case .Disabled:
-                            return callback(.Disabled)
-                        }
-                    } else {
-                        return callback(.Invalid)
-                    }
-                    
-                case .failure(_):
-                    ()
-                }
-            }
-    }
-    
-    
-    
-    /// A successful result here means pending (the session has been destroyed but the client still has data cached)
-    func signOut(afterRemoveSession : Bool = false) async -> Result<(), UserStateError>  {
-        
-        if !afterRemoveSession {
-            let status = try? await http.signout().get()
-            guard let status = status else { return .failure(.signOutError)}
-        }
-        
-        self.ws?.stop()
-        /*withAnimation {
-         state = .signedOut
-         }*/
-        // IMPORTANT: do not destroy the cache/session here. It'll cause the app to crash before it can transition to the welcome screen.
-        // The cache is destroyed in RevoltApp.swift:ApplicationSwitcher
-        
-        state = .signedOut
-        return .success(())
-    }
-    
-    
-    /// A workaround for the UserSettingStore finding out we're not authenticated, since not a main actor.
-    func setSignedOutState() {
-        withAnimation {
-            state = .signedOut
-        }
-    }
-    
-    func destroyCache() {
-        // In future this'll need to delete files too
-        path = []
-        
-        // MEMORY MANAGEMENT: Stop cleanup timer
-        memoryCleanupTimer?.invalidate()
-        memoryCleanupTimer = nil
-        
-        memoryMonitorTimer?.invalidate()
-        memoryMonitorTimer = nil
-        
-        // Cancel all pending saves
-        for workItem in saveWorkItems.values {
-            workItem.cancel()
-        }
-        saveWorkItems.removeAll()
-        
-        users.removeAll()
-        servers.removeAll()
-        channels.removeAll()
-        messages.removeAll()
-        members.removeAll()
-        emojis.removeAll()
-        dms.removeAll()
-        currentlyTyping.removeAll()
-        channelMessages.removeAll()
-        preloadedChannels.removeAll()
-        
-        currentUser = nil
-        currentSelection = .discover
-        currentChannel = .home
-        currentSessionId = nil
-        
-        userSettingsStore.isLoggingOut()
-        self.ws = nil
-    }
-    
     func promptForNotifications() async {
         let notificationsGranted = try? await UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .badge, .sound, .providesAppNotificationSettings])
         if notificationsGranted != nil && notificationsGranted! {
@@ -2138,25 +1111,6 @@ public class ViewState: ObservableObject {
         self.ws = ws
     }
     
-    func queueMessage(channel: String, replies: [Reply], content: String, attachments: [(Data, String)]) async {
-        var queue = self.queuedMessages[channel]
-        
-        if queue == nil {
-            queue = []
-            self.queuedMessages[channel] = queue
-        }
-        
-        let nonce = UUID().uuidString
-        
-        let r: [Revolt.ApiReply] = replies.map { reply in
-            Revolt.ApiReply(id: reply.message.id, mention: reply.mention)
-        }
-        
-        queue?.append(QueuedMessage(nonce: nonce, replies: r, content: content, author: currentUser?.id ?? "", channel: channel, timestamp: Date(), hasAttachments: !attachments.isEmpty, attachmentData: attachments))
-        
-        let _ = await http.sendMessage(channel: channel, replies: r, content: content, attachments: attachments, nonce: nonce)
-    }
-    
     /// Temporarily suspends WebSocket connection to reduce network conflicts when opening external URLs
     func temporarilySuspendWebSocket() {
         // print("🔌 Temporarily suspending WebSocket to prevent network conflicts")
@@ -2221,1190 +1175,12 @@ public class ViewState: ObservableObject {
         }
     }
     
-    private func processEvent(_ event: WsMessage) async {
-        switch event {
-        case .ready(let event):
-            // print("🚀 VIEWSTATE: Processing READY event")
-            // print("   - Users: \(event.users.count)")
-            // print("   - Servers: \(event.servers.count)")
-            // print("   - Channels: \(event.channels.count)")
-            // print("   - Members: \(event.members.count)")
-            // print("   - Emojis: \(event.emojis.count)")
-            
-            // CRITICAL FIX: Preserve current channel and selection during ready event
-            let savedCurrentChannel = currentChannel
-            let savedCurrentSelection = currentSelection
-            // print("💾 READY: Saving current state - channel: \(savedCurrentChannel), selection: \(savedCurrentSelection)")
-            
-            // CRITICAL FIX: Don't clear servers/channels/users completely - merge with existing data
-            // print("🔄 READY: Merging servers with existing data (current: \(servers.count), incoming: \(event.servers.count))")
-            
-            // Only clear messages as they should be fresh from server
-            messages.removeAll()
-            channelMessages.removeAll()
-            
-            // For users, channels, and servers: merge instead of clearing completely
-            // This preserves any data loaded from UserDefaults
-            // print("🔄 READY: Preserving existing servers/channels/users, will merge with server data")
-            
-            // MEMORY FIX: Extract only needed data and process immediately
-            // This allows the large event object to be released from memory
-            let neededData = extractNeededDataFromReadyEvent(event)
-            
-            // Process the extracted data
-            await processReadyData(neededData)
-            
-            // CRITICAL FIX: Restore saved state after ready event processing
-            // print("🔄 READY: Restoring saved state - channel: \(savedCurrentChannel), selection: \(savedCurrentSelection)")
-            currentChannel = savedCurrentChannel
-            currentSelection = savedCurrentSelection
-            
-            // If the saved channel is from a server, make sure that server's channels are loaded
-            if case .channel(let channelId) = savedCurrentChannel {
-                if let restoredChannel = allEventChannels[channelId] {
-                    // Make sure the channel is in active channels
-                    channels[channelId] = restoredChannel
-                    
-                    if let serverId = restoredChannel.server {
-                        // Make sure we're in the right selection and server channels are loaded
-                        if savedCurrentSelection != .server(serverId) {
-                            // print("🔄 READY: Correcting selection to server \(serverId) for channel \(channelId)")
-                            currentSelection = .server(serverId)
-                        }
-                        loadServerChannels(serverId: serverId)
-                    }
-                    // print("✅ READY: Successfully restored channel \(channelId)")
-                } else {
-                    // print("⚠️ READY: Could not restore channel \(channelId) - not found in stored channels")
-                }
-            }
-            
-            // CONDITIONAL: Only preload after Ready event if automatic preloading is enabled
-            if self.enableAutomaticPreloading {
-                // PRELOAD: Trigger preload of important channels after Ready event
-                Task {
-                    await self.preloadImportantChannels()
-                }
-                print("🚀 PRELOAD_ENABLED: Started automatic preloading after Ready event")
-            } else {
-                print("📵 PRELOAD_DISABLED: Skipped automatic preloading after Ready event")
-            }
-            
-            // CLEANUP: Clean up stale unreads after Ready event
-            // This ensures that unreads for deleted channels are removed after server sync
-            Task {
-                await MainActor.run {
-                    self.cleanupStaleUnreads()
-                    print("🧹 Cleaned up stale unreads after Ready event")
-                }
-            }
-
-        case .message(let m):
-            // print("📥 VIEWSTATE: Processing new message - id: \(m.id), channel: \(m.channel)")
-            // print("📥 VIEWSTATE: Current messages count BEFORE: \(messages.count)")
-            
-            if let user = m.user {
-                // CRITICAL FIX: Always add/update message authors to prevent black messages
-                users[user.id] = user
-                // CRITICAL FIX: Also store in allEventUsers for permanent access
-                allEventUsers[user.id] = user
-                // print("📥 VIEWSTATE: Added/updated user \(user.username) for message author to both dictionaries")
-            } else {
-                // CRITICAL FIX: If user data not provided, try to load from stored data or create placeholder
-                if users[m.author] == nil {
-                    if let storedUser = allEventUsers[m.author] {
-                        users[m.author] = storedUser
-                        // print("📥 VIEWSTATE: Loaded message author \(storedUser.username) from stored data")
-                    } else {
-                        // Create placeholder to prevent black messages
-                        let placeholderUser = Types.User(
-                            id: m.author,
-                            username: "Unknown User",
-                            discriminator: "0000",
-                            relationship: .None
-                        )
-                        users[m.author] = placeholderUser
-                        allEventUsers[m.author] = placeholderUser
-                        // print("⚠️ VIEWSTATE: Created placeholder for missing message author: \(m.author)")
-                    }
-                }
-            }
-            
-            if let member = m.member {
-                members[member.id.server]?[member.id.user] = member
-            }
-            
-            let userMentioned = m.mentions?.contains(where: {
-                $0 == currentUser?.id
-            }) ?? false
-            
-            // Check if message is from current user
-            let isFromCurrentUser = m.author == currentUser?.id
-                        
-            if let unread = unreads[m.channel]{
-                // Don't update unread for messages sent by the current user
-                if !isFromCurrentUser {
-                    // Update last_id for messages from other users
-                    // This ensures unread count properly reflects new messages
-                    unreads[m.channel]?.last_id = m.id
-                    
-                    if userMentioned {
-                        if unreads[m.channel]?.mentions != nil {
-                            unreads[m.channel]?.mentions?.append(m.id)
-                        } else {
-                            unreads[m.channel]!.mentions = [m.id]
-                        }
-                    }
-                }
-            } else if !isFromCurrentUser {
-                // Only create unread entry for messages from other users
-                unreads[m.channel] = .init(id: .init(channel: m.channel, user: currentUser?.id ?? ""),
-                                           last_id: m.id,
-                                           mentions: userMentioned ? [m.id]:[])
-            }
-            
-            // Check if message already exists
-            if messages[m.id] != nil {
-                // print("⚠️ VIEWSTATE: Message \(m.id) already exists, updating")
-            }
-            
-            messages[m.id] = m
-            
-            // Check if this message matches a queued message and clean it up
-            if let channelQueuedMessages = queuedMessages[m.channel],
-               let queuedIndex = channelQueuedMessages.firstIndex(where: { queued in
-                   // Match by content, author, and channel for safety
-                   return queued.content == m.content && 
-                          queued.author == m.author && 
-                          queued.channel == m.channel
-               }) {
-                let queuedMessage = channelQueuedMessages[queuedIndex]
-                print("📥 VIEWSTATE: Found matching queued message, cleaning up nonce: \(queuedMessage.nonce)")
-                
-                // Remove the temporary message from messages dictionary (if it exists)
-                messages.removeValue(forKey: queuedMessage.nonce)
-                
-                // For messages without attachments: Replace nonce with real ID in channel messages
-                // For messages with attachments: Add to channel messages for the first time
-                if let nonceMsgIndex = channelMessages[m.channel]?.firstIndex(of: queuedMessage.nonce) {
-                    // This was an optimistic message (no attachments), replace it
-                    channelMessages[m.channel]?[nonceMsgIndex] = m.id
-                    print("📥 VIEWSTATE: Replaced optimistic nonce \(queuedMessage.nonce) with real ID \(m.id)")
-                } else if queuedMessage.hasAttachments {
-                    // This was an attachment message (not shown optimistically), add it now
-                    if channelMessages[m.channel] == nil {
-                        channelMessages[m.channel] = []
-                    }
-                    channelMessages[m.channel]?.append(m.id)
-                    print("📥 VIEWSTATE: Added attachment message \(m.id) to channel messages for first time")
-                }
-                
-                // Remove from queued messages for this channel
-                queuedMessages[m.channel]?.remove(at: queuedIndex)
-                if queuedMessages[m.channel]?.isEmpty == true {
-                    queuedMessages.removeValue(forKey: m.channel)
-                }
-                print("📥 VIEWSTATE: Removed queued message from channel \(m.channel)")
-            } else {
-                // Check channel messages array
-                if channelMessages[m.channel] == nil {
-                    // print("📥 VIEWSTATE: Creating new channelMessages array for channel \(m.channel)")
-                    channelMessages[m.channel] = []
-                }
-                
-                // MEMORY FIX: Check if message already exists in channel to avoid duplicates
-                if !(channelMessages[m.channel]?.contains(m.id) ?? false) {
-                    channelMessages[m.channel]?.append(m.id)
-                } else {
-                    // print("⚠️ VIEWSTATE: Message \(m.id) already exists in channelMessages, skipping append")
-                }
-            }
-            
-            let channelMessagesAfter = channelMessages[m.channel]?.count ?? 0
-            
-            // print("📥 VIEWSTATE: Channel messages count - before: \(channelMessagesBefore), after: \(channelMessagesAfter)")
-            // print("📥 VIEWSTATE: Total messages count AFTER: \(messages.count)")
-            // print("📥 VIEWSTATE: Total channel message arrays: \(channelMessages.count)")
-            
-            // Log memory info
-            let totalChannelMessages = channelMessages.values.reduce(0) { $0 + $1.count }
-            // print("📥 VIEWSTATE: Total messages across all channels: \(totalChannelMessages)")
-            
-            // DISABLED: MEMORY MANAGEMENT: Proactive cleanup
-            // checkAndCleanupIfNeeded() - Disabled to prevent black messages
-            
-            NotificationCenter.default.post(name: NSNotification.Name("NewMessagesReceived"), object: nil)
-            
-            if let index = dms.firstIndex(where: { $0.id == m.channel }) {
-                let dmChannel = dms.remove(at: index)
-
-                let updatedDM: Channel
-                switch dmChannel {
-                    case .dm_channel(var c):
-                        c.last_message_id = m.id
-                        updatedDM = .dm_channel(c)
-                    case .group_dm_channel(var c):
-                        c.last_message_id = m.id
-                        updatedDM = .group_dm_channel(c)
-                    default:
-                        updatedDM = dmChannel
-                }
-
-                dms.insert(updatedDM, at: 0)
-                
-                // FIX: Ensure DM list state is maintained
-                if isDmListInitialized && currentSelection == .dms {
-                    // When a DM moves to top, ensure we maintain the loaded batches
-                    // because this change might affect the order
-                    let channelIdIndex = allDmChannelIds.firstIndex(of: m.channel)
-                    if let channelIdIndex = channelIdIndex {
-                        allDmChannelIds.remove(at: channelIdIndex)
-                        allDmChannelIds.insert(m.channel, at: 0)
-                    }
-                }
-            }
-            
-            if var existing = channels[m.channel] {
-                switch existing {
-                case .dm_channel(var c):
-                    c.last_message_id = m.id
-                    channels[m.channel] = .dm_channel(c)
-                case .group_dm_channel(var c):
-                    c.last_message_id = m.id
-                    channels[m.channel] = .group_dm_channel(c)
-                case .text_channel(var c):
-                    c.last_message_id = m.id
-                    channels[m.channel] = .text_channel(c)
-                default:
-                    break
-                }
-            }
-            
-            
-        case .message_update(let event):
-            let message = messages[event.id]
-            
-            if var message = message {
-                message.edited = event.data.edited
-                
-                
-                if let content = event.data.content {
-                    message.content = content
-                }
-                
-                messages[event.id] = message
-            }
-            
-        case .authenticated:
-            print("authenticated")
-            
-        case .invalid_session:
-            Task {
-                await self.signOut()
-            }
-            
-        case .logout:
-            Task {
-                await self.signOut(afterRemoveSession: true)
-            }
-        case .channel_start_typing(let e):
-            var typing = currentlyTyping[e.id] ?? []
-            typing.append(e.user)
-            
-            currentlyTyping[e.id] = typing
-            
-        case .channel_stop_typing(let e):
-            currentlyTyping[e.id]?.removeAll(where: { $0 == e.user })
-            
-        case .message_delete(let e):
-            if var channel = channelMessages[e.channel] {
-                if let index = channel.firstIndex(of: e.id) {
-                    channel.remove(at: index)
-                    channelMessages[e.channel] = channel
-                }
-            }
-            
-        case .channel_ack(let e):
-            unreads[e.id]?.last_id = e.message_id
-            unreads[e.id]?.mentions?.removeAll { $0 <= e.message_id }
-            
-        case .message_react(let e):
-            if var message = messages[e.id] {
-                var reactions = message.reactions ?? [:]
-                var users = reactions[e.emoji_id] ?? []
-                
-                // Check if user is not already in the reaction list to avoid duplicates
-                if !users.contains(e.user_id) {
-                    users.append(e.user_id)
-                    reactions[e.emoji_id] = users
-                    message.reactions = reactions
-                    messages[e.id] = message
-                    
-                    // print("🔥 VIEWSTATE: Added reaction \(e.emoji_id) from user \(e.user_id) to message \(e.id) in channel \(e.channel_id)")
-                    
-                    // Post notification to update UI
-                    DispatchQueue.main.async {
-                        NotificationCenter.default.post(
-                            name: NSNotification.Name("MessagesDidChange"), 
-                            object: ["channelId": e.channel_id, "messageId": e.id, "type": "reaction_added"]
-                        )
-                    }
-                } else {
-                    // print("🔥 VIEWSTATE: User \(e.user_id) already reacted with \(e.emoji_id) on message \(e.id)")
-                }
-            } else {
-                // print("🔥 VIEWSTATE: Message \(e.id) not found for reaction add")
-            }
-            
-        case .message_unreact(let e):
-            if var message = messages[e.id] {
-                if var reactions = message.reactions {
-                    if var users = reactions[e.emoji_id] {
-                        users.removeAll { $0 == e.user_id }
-                        
-                        if users.isEmpty {
-                            reactions.removeValue(forKey: e.emoji_id)
-                        } else {
-                            reactions[e.emoji_id] = users
-                        }
-                        message.reactions = reactions
-                        messages[e.id] = message
-                        
-                        // print("🔥 VIEWSTATE: Removed reaction \(e.emoji_id) from user \(e.user_id) on message \(e.id) in channel \(e.channel_id)")
-                        
-                        // Post notification to update UI
-                        DispatchQueue.main.async {
-                            NotificationCenter.default.post(
-                                name: NSNotification.Name("MessagesDidChange"), 
-                                object: ["channelId": e.channel_id, "messageId": e.id, "type": "reaction_removed"]
-                            )
-                        }
-                    } else {
-                        // print("🔥 VIEWSTATE: No users found for emoji \(e.emoji_id) on message \(e.id)")
-                    }
-                } else {
-                    // print("🔥 VIEWSTATE: No reactions found on message \(e.id)")
-                }
-            } else {
-                // print("🔥 VIEWSTATE: Message \(e.id) not found for reaction remove")
-            }
-        case .message_append(let e):
-            if var message = messages[e.id] {
-                var embeds = message.embeds ?? []
-                embeds.append(e.append)
-                message.embeds = embeds
-                messages[e.id] = message
-            }
-        case .user_update(let e):
-            updateUser(with: e)
-        case .server_create(let e):
-            self.servers[e.id] = e.server
-            for channel in e.channels {
-                self.channels[channel.id] = channel
-                self.channelMessages[channel.id] = []
-            }
-            
-        case .server_delete(let e):
-            if case .server(let string) = currentSelection {
-                if string == e.id {
-                    self.path = .init()
-                    self.selectDms()
-                }
-            }
-            
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.75) {
-                self.servers.removeValue(forKey: e.id)
-            }
-            
-            
-        case .server_update(let e):
-            
-            if let tmpServer = self.servers[e.id] {
-                
-                var t = tmpServer
-                
-                if let icon = e.data?.icon {
-                    t.icon = icon
-                }
-                
-                if let name = e.data?.name  {
-                    t.name = name
-                }
-                
-                if let description = e.data?.description {
-                    t.description = description
-                }
-                
-                if let banner = e.data?.banner {
-                    t.banner = banner
-                }
-                
-                if let systemMessage = e.data?.system_messages {
-                    t.system_messages = systemMessage
-                }
-                
-                if let categories = e.data?.categories {
-                    t.categories = categories
-                }
-                
-                if let default_permissions = e.data?.default_permissions {
-                    t.default_permissions = default_permissions
-                }
-                
-                if let owner = e.data?.owner {
-                    t.owner = owner
-                }
-                
-                if let nsfw = e.data?.nsfw {
-                    t.nsfw = nsfw
-                }
-                
-                
-                if e.clear?.contains(ServerEdit.Remove.icon) == true {
-                    t.icon = nil
-                }
-                
-                if e.clear?.contains(ServerEdit.Remove.banner) == true {
-                    t.banner = nil
-                }
-                
-                self.servers[e.id] = t
-            }
-            
-        case .channel_create(let channel):
-            // Store the channel in our event channels for lazy loading
-            allEventChannels[channel.id] = channel
-            
-            // Handle different channel types
-            switch channel {
-            case .dm_channel(_):
-                // DMs are always loaded immediately
-                self.channels[channel.id] = channel
-                self.channelMessages[channel.id] = []
-                self.dms.insert(channel, at: 0)
-                // print("📥 VIEWSTATE: Added new DM channel \(channel.id) immediately")
-                
-            case .group_dm_channel(_):
-                // Group DMs are always loaded immediately
-                self.channels[channel.id] = channel
-                self.channelMessages[channel.id] = []
-                self.dms.insert(channel, at: 0)
-                // print("📥 VIEWSTATE: Added new Group DM channel \(channel.id) immediately")
-                
-            case .text_channel(let textChannel):
-                // Server channels: only load if server is currently active
-                if case .server(let currentServerId) = currentSelection, 
-                   currentServerId == textChannel.server {
-                    // Load immediately if this server is active
-                    self.channels[channel.id] = channel
-                    self.channelMessages[channel.id] = []
-                    // print("📥 VIEWSTATE: Added new text channel \(channel.id) immediately (server active)")
-                } else {
-                    // Just store for lazy loading later
-                    // print("🔄 LAZY_CHANNEL: Stored new text channel \(channel.id) for lazy loading")
-                }
-                
-                // Update server's channel list
-                if let serverId = channel.server {
-                    self.servers[serverId]?.channels.append(channel.id)
-                }
-                
-            case .voice_channel(let voiceChannel):
-                // Voice channels: only load if server is currently active
-                if case .server(let currentServerId) = currentSelection,
-                   currentServerId == voiceChannel.server {
-                    // Load immediately if this server is active
-                    self.channels[channel.id] = channel
-                    // print("📥 VIEWSTATE: Added new voice channel \(channel.id) immediately (server active)")
-                } else {
-                    // Just store for lazy loading later
-                    // print("🔄 LAZY_CHANNEL: Stored new voice channel \(channel.id) for lazy loading")
-                }
-                
-                // Update server's channel list
-                if let serverId = channel.server {
-                    self.servers[serverId]?.channels.append(channel.id)
-                }
-                
-            default:
-                // Other channel types - store in event channels
-                print("📥 VIEWSTATE: Stored unknown channel type \(channel.id)")
-            }
-            
-            // Update app badge count when new channel is created
-            // This ensures unread messages in the new channel are counted
-            updateAppBadgeCount()
-            
-        case .channel_update(let e):
-            
-            if let index = self.dms.firstIndex(where: { $0.id == e.id }),
-               case .group_dm_channel(var groupDMChannel) = self.dms[index] {
-                
-                
-                if let name = e.data?.name {
-                    groupDMChannel.name = name
-                }
-                
-                if let icon = e.data?.icon {
-                    groupDMChannel.icon = icon
-                }
-                
-                if let description = e.data?.description {
-                    groupDMChannel.description = description
-                }
-                
-                if let nsfw = e.data?.nsfw {
-                    groupDMChannel.nsfw = nsfw
-                }
-                
-                if let permission = e.data?.permissions {
-                    groupDMChannel.permissions = permission
-                }
-                
-                if let owner = e.data?.owner {
-                    groupDMChannel.owner = owner
-                }
-                
-                if e.clear?.contains(.icon) == true {
-                    groupDMChannel.icon = nil
-                }
-                
-                if e.clear?.contains(.description) == true {
-                    groupDMChannel.description = nil
-                }
-                
-                
-                
-                self.dms[index] = .group_dm_channel(groupDMChannel)
-                
-            } else if let index = self.dms.firstIndex(where: { $0.id == e.id }),
-                      case .dm_channel(let dmChannel) = self.dms[index] {
-                
-                //TODO
-                
-                self.dms[index] = .dm_channel(dmChannel)
-                
-            }
-            
-            if let channel = self.channels[e.id] {
-                
-                
-                if case .group_dm_channel(var t) = channel {
-                    if let name = e.data?.name {
-                        t.name = name
-                    }
-                    
-                    if let icon = e.data?.icon {
-                        t.icon = icon
-                    }
-                    
-                    
-                    if let description = e.data?.description {
-                        t.description = description
-                    }
-                    
-                    if let nsfw = e.data?.nsfw {
-                        t.nsfw = nsfw
-                    }
-                    
-                    if let permission = e.data?.permissions {
-                        t.permissions = permission
-                    }
-                    
-                    if let owner = e.data?.owner {
-                        t.owner = owner
-                    }
-                    
-                    
-                    if e.clear?.contains(.icon) == true {
-                        t.icon = nil
-                    }
-                    
-                    if e.clear?.contains(.description) == true {
-                        t.description = nil
-                    }
-                    
-                    
-                    self.channels[e.id] = .group_dm_channel(t)
-                    
-                    
-                } else if case .text_channel(var t) = channel {
-                    if let name = e.data?.name {
-                        t.name = name
-                    }
-                    
-                    if let icon = e.data?.icon {
-                        t.icon = icon
-                    }
-                    
-                    if let description = e.data?.description {
-                        t.description = description
-                    }
-                    
-                    if let nsfw = e.data?.nsfw {
-                        t.nsfw = nsfw
-                    }
-                    
-                    if let default_permissions = e.data?.default_permissions {
-                        t.default_permissions = default_permissions
-                    }
-                    
-                    if let newRolePermissions = e.data?.role_permissions {
-                        if t.role_permissions == nil {
-                            t.role_permissions = newRolePermissions
-                        } else {
-                            for (roleId, permission) in newRolePermissions {
-                                t.role_permissions?[roleId] = permission
-                            }
-                        }
-                    }
-                    
-                    if e.clear?.contains(.icon) == true {
-                        t.icon = nil
-                    }
-                    
-                    if e.clear?.contains(.description) == true {
-                        t.description = nil
-                    }
-                    
-                    self.channels[e.id] = .text_channel(t)
-                    
-                }
-            }
-            
-        case .channel_delete(let e):
-            self.deleteChannel(channelId: e.id)
-            
-        case .channel_group_leave(let e):
-            if e.user == currentUser?.id {
-                deleteChannel(channelId: e.id)
-            } else {
-                
-                if case .group_dm_channel(var channel) = self.channels[e.id] {
-                    channel.recipients.removeAll { $0 == e.user }
-                    self.channels[e.id] = .group_dm_channel(channel)
-                    if let index = dms.firstIndex(where: { $0.id == e.id }) {
-                        dms[index] = .group_dm_channel(channel)
-                    }
-                } else {
-                    //Todo
-                }
-                
-            }
-            
-        case .channel_group_join(let e):
-            if case .group_dm_channel(var channel) = self.channels[e.id] {
-                channel.recipients.append(e.user)
-                self.channels[e.id] = .group_dm_channel(channel)
-                if let index = dms.firstIndex(where: { $0.id == e.id }) {
-                    dms[index] = .group_dm_channel(channel)
-                }
-                
-                //TOOD
-                //fetch user
-                let response = await self.http.fetchUser(user: e.user)
-                switch response {
-                    case .success(let user):
-                        // MEMORY FIX: Only add users if we have space
-                        if self.users.count < self.maxUsersInMemory {
-                            self.users[user.id] = user
-                            // print("📥 VIEWSTATE: Added user \(user.id) during channel_group_join")
-                        }
-                        self.checkAndCleanupIfNeeded()
-                        
-                    case .failure(let error):
-                        print(error)
-                }
-                
-            } else {
-                //Todo other types channel
-            }
-            
-            // MEMORY MANAGEMENT: Cleanup after new user
-            checkAndCleanupIfNeeded()
-            
-            
-            
-        case .server_member_update(let e):
-            let serverId = e.id.server
-            let userId = e.id.user
-
-            guard var serverMembers = members[serverId], var member = serverMembers[userId] else {
-                return
-            }
-
-            // Apply updates only to non-nil fields
-            if let newNickname = e.data?.nickname {
-                member.nickname = newNickname
-            }
-            
-            if let newAvatar = e.data?.avatar {
-                member.avatar = newAvatar
-            }
-            
-            if let newRoles = e.data?.roles {
-                member.roles = newRoles
-            }
-            
-            if let newJoinedAt = e.data?.joined_at {
-                member.joined_at = newJoinedAt
-            }
-            
-            if let newTimeout = e.data?.timeout {
-                member.timeout = newTimeout
-            }
-
-            // Handle `clean` fields (removing values if specified)
-            for field in e.clear {
-                switch field {
-                case .nickname:
-                    member.nickname = nil
-                case .avatar:
-                    member.avatar = nil
-                case .roles:
-                    member.roles = nil
-                case .timeout:
-                    member.timeout = nil
-                }
-            }
-
-            // Update the local members dictionary
-            serverMembers[userId] = member
-            members[serverId] = serverMembers
-            
-        case .server_member_join(let e):
-
-            Task {
-                async let fetchedUser = self.http.fetchUser(user: e.user)
-                async let fetchedMember = self.http.fetchMember(server: e.id, member: e.user)
-
-                // Wait for both API calls to complete
-                let (userResult, memberResult) = await (fetchedUser, fetchedMember)
-                
-                
-                switch userResult {
-                    case .success(let user):
-                        // MEMORY FIX: Only add users if we have space
-                        if self.users.count < self.maxUsersInMemory {
-                            self.users[e.user] = user
-                            // print("📥 VIEWSTATE: Added user \(e.user) during server_member_join")
-                        }
-                        self.checkAndCleanupIfNeeded()
-                    case .failure(_):
-                         print("error fetching user")
-                }
-                
-                switch memberResult {
-                    case .success(let member):
-                        var serverMembers = self.members[e.id, default: [:]]
-                        serverMembers[e.user] = member
-                        self.members[e.id] = serverMembers
-                    case .failure(_):
-                         print("error fetching member")
-                }
-
-            }
-            
-        case .server_member_leave(let e):
-                guard var serverMembers = self.members[e.id] else {
-                    return
-                }
-                serverMembers.removeValue(forKey: e.user)
-                self.members[e.id] = serverMembers
-            
-        case .server_role_update(let e):
-                // Ensure the server exists
-                guard var server = self.servers[e.id] else {
-                    return
-                }
-                
-                // Ensure the roles dictionary exists
-                var serverRoles = server.roles ?? [:]
-                
-                // Check if the role already exists
-                var role = serverRoles[e.role_id] ?? Role(
-                    name: e.data.name ?? "New Role",
-                    permissions: e.data.permissions ?? Overwrite(a: .none, d: .none),
-                    colour: e.data.colour,
-                    hoist: e.data.hoist,
-                    rank: e.data.rank ?? 0
-                )
-
-                // Update fields if they exist in the event
-                if let name = e.data.name {
-                    role.name = name
-                }
-                if let permissions = e.data.permissions as Overwrite? {
-                    role.permissions = permissions
-                }
-                if let colour = e.data.colour {
-                    role.colour = colour
-                }
-                if let hoist = e.data.hoist {
-                    role.hoist = hoist
-                }
-                if let rank = e.data.rank {
-                    role.rank = rank
-                }
-
-                // Remove fields specified in `clear`
-                for field in e.clear {
-                    switch field {
-                    case .colour:
-                        role.colour = nil
-                    }
-                }
-
-                // Save the updated role
-                serverRoles[e.role_id] = role
-                server.roles = serverRoles
-                self.servers[e.id] = server
-            
-            
-        case .server_role_delete(let e):
-                // Ensure the server exists
-                guard var server = self.servers[e.id] else {
-                    return
-                }
-                
-                // Ensure the roles dictionary exists
-                guard var serverRoles = server.roles else {
-                    return
-                }
-            
-                serverRoles.removeValue(forKey: e.role_id)
-               
-                // Update the server's roles
-                server.roles = serverRoles
-                self.servers[e.id] = server
-            
-        case .user_relationship(let event):
-            updateUserRelationship(with: event)
-            
-        case .user_setting_update(let event):
-            
-            if let update = event.update {
-                self.userSettingsStore.storeFetchData(settingsValues: update)
-                
-                if update["ordering"] != nil {
-                    DispatchQueue.main.async {
-                        self.applyServerOrdering()
-                    }
-                }
-            }
-            
-        }
-        
-    }
-    
-
-
-    
-    private func processUsers(_ eventUsers: [Types.User]) {
-        // print("🚀 VIEWSTATE: Processing \(eventUsers.count) users from WebSocket")
-        // print("🚀 VIEWSTATE: Existing users count: \(users.count)")
-        
-        // Store ALL users for lazy loading (this is our data source)
-        allEventUsers = Dictionary(uniqueKeysWithValues: eventUsers.map { ($0.id, $0) })
-        // print("🔄 LAZY_USER: Stored \(allEventUsers.count) users for lazy loading")
-        
-        // CRITICAL FIX: Don't clear existing users - merge instead
-        // Keep existing users and add/update new ones from WebSocket
-        
-        var addedCount = 0
-        var updatedCount = 0
-        var currentUserFound = false
-        
-        // 1. Update/Add current user
-        for user in eventUsers {
-            if user.relationship == .User {
-                currentUser = user
-                if users[user.id] == nil {
-                    users[user.id] = user
-                    addedCount += 1
-                    // print("🚀 VIEWSTATE: Added current user: \(user.id)")
-                } else {
-                    users[user.id] = user
-                    updatedCount += 1
-                    // print("🚀 VIEWSTATE: Updated current user: \(user.id)")
-                }
-                currentUserFound = true
-                break
-            }
-        }
-        
-        // 2. Update/Add friends (always important)
-        for user in eventUsers {
-            if addedCount >= 50 {
-                break
-            }
-            
-            if user.relationship == .Friend {
-                if users[user.id] == nil {
-                    users[user.id] = user
-                    addedCount += 1
-                    // print("🚀 VIEWSTATE: Added friend: \(user.id)")
-                } else {
-                    users[user.id] = user
-                    updatedCount += 1
-                    // print("🚀 VIEWSTATE: Updated friend: \(user.id)")
-                }
-            }
-        }
-        
-        // 3. Add users needed for visible DMs only (lazy approach)
-        // Note: This will be called later in processDMs after allDmChannelIds is set
-        
-        // print("🚀 VIEWSTATE: FINAL USER COUNT: \(users.count) (added: \(addedCount), updated: \(updatedCount)) out of \(eventUsers.count) total")
-        // print("🔄 LAZY_USER: Remaining users will be loaded on-demand")
-        
-        if !currentUserFound {
-            // print("⚠️ VIEWSTATE: Current user not found in event users!")
-        }
-    }
-    
-    // Load users needed for currently visible DMs
-    private func loadUsersForVisibleDms(from userDict: [String: Types.User], maxCount: Int) {
-        var loadedCount = 0
-        
-        // Get IDs for first batch of DMs that will be visible
-        let visibleDmIds = Array(allDmChannelIds.prefix(dmBatchSize))
-        
-        for dmId in visibleDmIds {
-            if loadedCount >= maxCount {
-                break
-            }
-            
-            if let channel = channels[dmId] {
-                var recipientIds: [String] = []
-                
-                switch channel {
-                case .dm_channel(let dm):
-                    recipientIds = dm.recipients
-                case .group_dm_channel(let group):
-                    recipientIds = group.recipients
-                default:
-                    continue
-                }
-                
-                // Load users for this DM
-                for userId in recipientIds {
-                    if loadedCount >= maxCount {
-                        break
-                    }
-                    
-                    if users[userId] == nil, let user = userDict[userId] {
-                        users[userId] = user
-                        loadedCount += 1
-                        // print("🔄 LAZY_USER: Loaded DM participant: \(userId)")
-                    }
-                }
-            }
-        }
-        
-        // print("🔄 LAZY_USER: Loaded \(loadedCount) users for visible DMs")
-    }
-    
     // Store user data for lazy loading
     var allEventUsers: [String: Types.User] = [:]
     
     // LAZY LOADING: Server channel management
     var allEventChannels: [String: Channel] = [:] // Store all channels for lazy loading
     var loadedServerChannels: Set<String> = [] // Track which servers have loaded channels
-    
-    // Load users for the first batch of DMs (called during processDMs)
-    private func loadUsersForFirstDmBatch() {
-        var loadedCount = 0
-        let maxUsersToLoad = 50 // Limit to prevent memory issues
-        
-        // Get IDs for first batch of DMs that will be visible
-        let visibleDmIds = Array(allDmChannelIds.prefix(dmBatchSize))
-        
-        for dmId in visibleDmIds {
-            if loadedCount >= maxUsersToLoad {
-                break
-            }
-            
-            if let channel = channels[dmId] {
-                var recipientIds: [String] = []
-                
-                switch channel {
-                case .dm_channel(let dm):
-                    recipientIds = dm.recipients
-                case .group_dm_channel(let group):
-                    recipientIds = group.recipients
-                default:
-                    continue
-                }
-                
-                // Load actual users from stored event data
-                for userId in recipientIds {
-                    if loadedCount >= maxUsersToLoad {
-                        break
-                    }
-                    
-                    if users[userId] == nil {
-                        if let actualUser = allEventUsers[userId] {
-                            // Load the real user data
-                            users[userId] = actualUser
-                            loadedCount += 1
-                            // print("🔄 LAZY_USER: Loaded actual user \(actualUser.username) for DM participant: \(userId)")
-                        } else {
-                            // Create placeholder only if we can't find the real user
-                            let placeholderUser = Types.User(
-                                id: userId,
-                                username: "Unknown User",
-                                discriminator: "0000",
-                                relationship: .None
-                            )
-                            users[userId] = placeholderUser
-                            loadedCount += 1
-                            // print("⚠️ LAZY_USER: Created placeholder for missing user: \(userId)")
-                        }
-                    }
-                }
-            }
-        }
-        
-        // print("🔄 LAZY_USER: Loaded \(loadedCount) users for first DM batch")
-    }
-    
-    // Load users on-demand when a new DM batch is loaded
-    @MainActor
-    func loadUsersForDmBatch(_ batchIndex: Int) {
-        let startIndex = batchIndex * dmBatchSize
-        let endIndex = min(startIndex + dmBatchSize, allDmChannelIds.count)
-        
-        guard startIndex < allDmChannelIds.count else {
-            return
-        }
-        
-        let batchIds = Array(allDmChannelIds[startIndex..<endIndex])
-        var loadedCount = 0
-        var skippedCount = 0
-        let maxUsersToLoad = 10 // REDUCED to 10 users per batch for debugging
-        
-        for dmId in batchIds {
-            if loadedCount >= maxUsersToLoad {
-                break
-            }
-            
-            if let channel = channels[dmId] {
-                var recipientIds: [String] = []
-                
-                switch channel {
-                case .dm_channel(let dm):
-                    recipientIds = dm.recipients
-                case .group_dm_channel(let group):
-                    recipientIds = group.recipients
-                default:
-                    continue
-                }
-                
-                // Load actual users from stored event data
-                for userId in recipientIds {
-                    if loadedCount >= maxUsersToLoad {
-                        break
-                    }
-                    
-                    // DUPLICATE PREVENTION: Skip if user already exists
-                    if users[userId] != nil {
-                        skippedCount += 1
-                        continue
-                    }
-                    
-                    if let actualUser = allEventUsers[userId] {
-                        // Load the real user data
-                        users[userId] = actualUser
-                        loadedCount += 1
-                        // print("🔄 LAZY_USER: Loaded NEW user \(actualUser.username) for DM batch \(batchIndex)")
-                    } else {
-                        // print("⚠️ LAZY_USER: User \(userId) not found in event data for batch \(batchIndex)")
-                    }
-                }
-            }
-        }
-        
-        // print("🔄 LAZY_USER: Batch \(batchIndex) - Loaded \(loadedCount) NEW users, skipped \(skippedCount) existing users. Total users now: \(users.count)")
-    }
-    
-    // Load users for visible messages to prevent black messages
-    @MainActor
-    func loadUsersForVisibleMessages(channelId: String) {
-        guard let messageIds = channelMessages[channelId] else {
-            return
-        }
-        
-        var loadedUsers = 0
-        var missingUsers: [String] = []
-        
-        for messageId in messageIds {
-            if let message = messages[messageId] {
-                if users[message.author] == nil {
-                    missingUsers.append(message.author)
-                    
-                    // Try to load from event data first
-                    if let user = allEventUsers[message.author] {
-                        users[message.author] = user
-                        loadedUsers += 1
-                        // print("🔄 LAZY_USER: Loaded message author \(user.username) from event data")
-                    } else {
-                        // Create placeholder user to prevent black messages
-                        let placeholderUser = Types.User(
-                            id: message.author,
-                            username: "Unknown User",
-                            discriminator: "0000",
-                            relationship: .None
-                        )
-                        users[message.author] = placeholderUser
-                        loadedUsers += 1
-                        // print("⚠️ LAZY_USER: Created placeholder for missing user: \(message.author)")
-                    }
-                }
-            }
-        }
-        
-        if loadedUsers > 0 {
-            // print("🔄 LAZY_USER: Loaded \(loadedUsers) users for channel \(channelId), missing: \(missingUsers.count)")
-        }
-    }
-    
-    // CRITICAL FIX: Restore missing users from allEventUsers to prevent black messages
-    @MainActor
-    func restoreMissingUsersForMessages() {
-        var restoredCount = 0
-        var placeholderCount = 0
-        
-        // print("🔄 RESTORE_USERS: Starting restoration of missing users")
-        
-        // Check all messages in memory
-        for (messageId, message) in messages {
-            if users[message.author] == nil {
-                // Try to restore from allEventUsers
-                if let storedUser = allEventUsers[message.author] {
-                    users[message.author] = storedUser
-                    restoredCount += 1
-                    // print("🔄 RESTORE_USERS: Restored \(storedUser.username) for message \(messageId)")
-                } else {
-                    // Create placeholder as last resort
-                    let placeholderUser = Types.User(
-                        id: message.author,
-                        username: "Unknown User",
-                        discriminator: "0000",
-                        relationship: .None
-                    )
-                    users[message.author] = placeholderUser
-                    allEventUsers[message.author] = placeholderUser // Store for future use
-                    placeholderCount += 1
-                    // print("⚠️ RESTORE_USERS: Created placeholder for \(message.author) in message \(messageId)")
-                }
-            }
-        }
-        
-        if restoredCount > 0 || placeholderCount > 0 {
-            // print("🔄 RESTORE_USERS: Restoration complete - restored: \(restoredCount), placeholders: \(placeholderCount)")
-        }
-    }
     
     // EMERGENCY FIX: Force restore all users for specific channel to prevent black messages
     @MainActor
@@ -3572,138 +1348,6 @@ public class ViewState: ObservableObject {
         
         // print("🔄 LAZY_CHANNEL: Unloaded \(removedCount) channels for server \(serverId)")
         // print("🔄 LAZY_CHANNEL: Total active channels now: \(channels.count)")
-    }
-    
-    private func processMembers(_ eventMembers: [Member]) {
-        for member in eventMembers {
-            members[member.id.server]?[member.id.user] = member
-        }
-    }
-    
-    private func processDMs(channels: [Channel]) {
-        // LAZY LOADING: Store all DM IDs but only load the first batch
-        let dmChannels: [Channel] = channels.filter {
-            switch $0 {
-            case .dm_channel:
-                return true // Include both active and inactive DMs
-            case .group_dm_channel:
-                return true
-            default:
-                return false
-            }
-        }
-        
-        // print("🚀 VIEWSTATE: Processing \(dmChannels.count) DM channels with lazy loading")
-        
-        // Sort all DM channels
-        let sortedDmChannels = dmChannels.sorted { first, second in
-            let firstLast = first.last_message_id
-            let secondLast = second.last_message_id
-            
-            let firstUnreadLast = unreads[first.id]?.last_id
-            let secondUnreadLast = unreads[second.id]?.last_id
-            
-            let firstIsUnread = firstLast != nil && firstLast != firstUnreadLast
-            let secondIsUnread = secondLast != nil && secondLast != secondUnreadLast
-            
-            // Show unread DMs first
-            if firstIsUnread && !secondIsUnread {
-                return true
-            } else if !firstIsUnread && secondIsUnread {
-                return false
-            } else {
-                return (firstLast ?? "") > (secondLast ?? "")
-            }
-        }
-        
-        // Store all DM IDs in order
-        allDmChannelIds = sortedDmChannels.map { $0.id }
-        
-        // Load users for visible DMs after we have the sorted list
-        loadUsersForFirstDmBatch()
-        
-        // Simple lazy loading: start fresh
-        loadedDmBatches.removeAll()
-        dms.removeAll() // Clear existing DMs
-        loadDmBatch(0) // Load first batch
-        
-        isDmListInitialized = true
-        // print("🚀 VIEWSTATE: Stored \(allDmChannelIds.count) DM IDs, loaded first batch")
-    }
-    
-    // Load a specific batch of DMs
-    @MainActor
-    func loadDmBatch(_ batchIndex: Int) {
-        guard !isLoadingDmBatch else {
-            // print("🔄 LAZY_DM: Already loading, skipping batch \(batchIndex)")
-            return
-        }
-        
-        // DUPLICATE PREVENTION: Check if this batch is already loaded
-        if loadedDmBatches.contains(batchIndex) {
-            // print("🔄 LAZY_DM: Batch \(batchIndex) already loaded, skipping")
-            return
-        }
-        
-        let startIndex = batchIndex * dmBatchSize
-        let endIndex = min(startIndex + dmBatchSize, allDmChannelIds.count)
-        
-        guard startIndex < allDmChannelIds.count else {
-            return
-        }
-        
-        isLoadingDmBatch = true
-        
-        let memoryBefore = getCurrentMemoryUsage()
-        // print("🔄 LAZY_DM: Loading batch \(batchIndex) (DMs \(startIndex) to \(endIndex-1)) - Memory: \(memoryBefore)MB")
-        
-        // Get batch IDs to load
-        let batchIds = Array(allDmChannelIds[startIndex..<endIndex])
-        var newDms: [Channel] = []
-        
-        for dmId in batchIds {
-            if let channel = channels[dmId] {
-                newDms.append(channel)
-            }
-        }
-        
-        // FIXED: Rebuild the entire DMs list from allDmChannelIds to maintain correct order
-        // Mark this batch as loaded first
-        loadedDmBatches.insert(batchIndex)
-        
-        // Now rebuild the DMs list from all loaded batches in correct order
-        var rebuiltDms: [Channel] = []
-        var addedChannelIds = Set<String>() // Prevent duplicates
-        
-        for loadedBatch in loadedDmBatches.sorted() {
-            let batchStart = loadedBatch * dmBatchSize
-            let batchEnd = min(batchStart + dmBatchSize, allDmChannelIds.count)
-            
-            for i in batchStart..<batchEnd {
-                let channelId = allDmChannelIds[i]
-                if !addedChannelIds.contains(channelId), let channel = channels[channelId] {
-                    rebuiltDms.append(channel)
-                    addedChannelIds.insert(channelId)
-                }
-            }
-        }
-        
-        // Replace the entire DMs list with the rebuilt one
-        dms = rebuiltDms
-        
-        // Load users for this batch
-        loadUsersForDmBatch(batchIndex)
-        
-        // Simple memory protection: if too many batches, stop loading
-        if loadedDmBatches.count >= maxLoadedBatches {
-            // print("⚠️ LAZY_DM: Reached max batches limit (\(maxLoadedBatches)). No more loading.")
-        }
-        
-        let memoryAfter = getCurrentMemoryUsage()
-        let memoryDiff = memoryAfter - memoryBefore
-        
-        isLoadingDmBatch = false
-        // print("🔄 LAZY_DM: Loaded batch \(batchIndex), total DMs: \(dms.count), Memory: \(memoryBefore)MB → \(memoryAfter)MB (\(memoryDiff > 0 ? "+" : "")\(memoryDiff)MB)")
     }
     
     /* DISABLED: Virtual Scrolling functions
@@ -3912,20 +1556,6 @@ public class ViewState: ObservableObject {
     }
     */ // END DISABLED Virtual Scrolling functions
     
-    // Load next batch when user scrolls to bottom
-    @MainActor
-    func loadMoreDmsIfNeeded() {
-        // Find the highest loaded batch and load the next one
-        let maxLoadedBatch = loadedDmBatches.max() ?? -1
-        let nextBatchIndex = maxLoadedBatch + 1
-        let totalBatches = (allDmChannelIds.count + dmBatchSize - 1) / dmBatchSize
-        
-        // Check if we haven't reached the limit and there are more batches
-        if nextBatchIndex < totalBatches && loadedDmBatches.count < maxLoadedBatches {
-            loadDmBatch(nextBatchIndex)
-        }
-    }
-    
     // Check if there are more DMs to load
     var hasMoreDmsToLoad: Bool {
         let nextBatchIndex = loadedDmBatches.count
@@ -3961,40 +1591,6 @@ public class ViewState: ObservableObject {
         
         // CRITICAL FIX: Check if we have gaps in loaded batches and fill them
         ensureNoBatchGaps()
-    }
-    
-    // Ensure there are no gaps in loaded batches that could cause missing DMs
-    @MainActor
-    func ensureNoBatchGaps() {
-        guard !loadedDmBatches.isEmpty else { return }
-        
-        let sortedBatches = loadedDmBatches.sorted()
-        let minBatch = sortedBatches.first!
-        let maxBatch = sortedBatches.last!
-        
-        // Fill any gaps between min and max loaded batches
-        for batchIndex in minBatch...maxBatch {
-            if !loadedDmBatches.contains(batchIndex) {
-                // print("🔄 LAZY_DM: Found gap at batch \(batchIndex), loading to fill gap")
-                loadDmBatch(batchIndex)
-            }
-        }
-    }
-    
-    // Reset and reload DMs list (useful for fixing display issues)
-    @MainActor
-    func resetAndReloadDms() {
-        // print("🔄 DM_RESET: Resetting and reloading DMs list")
-        
-        // Clear current state
-        dms.removeAll()
-        loadedDmBatches.removeAll()
-        isLoadingDmBatch = false
-        
-        // Reload first batch
-        if !allDmChannelIds.isEmpty {
-            loadDmBatch(0)
-        }
     }
     
     // Ensure all visible batches are loaded based on current visible range
@@ -4308,156 +1904,6 @@ public class ViewState: ObservableObject {
         
     }
     
-    
-    func getUnreadCountFor(channel: Channel) -> UnreadCount? {
-        /*if let unread = unreads[channel.id] {
-         if let mentions = unread.mentions {
-         return .mentions(formattedMentionCount(mentions.count))
-         }
-         
-         if let last_unread_id = unread.last_id, let last_message_id = channel.last_message_id {
-         if last_unread_id < last_message_id {
-         return .unread
-         }
-         }
-         }
-         
-         return nil*/
-        
-        guard let unread = unreads[channel.id] else {
-            return nil
-        }
-        
-        let hasMentions = (unread.mentions != nil && unread.mentions?.count ?? 0 > 0)
-        
-        
-        
-        let hasUnread: Bool = {
-            if let lastUnreadId = unread.last_id, let lastMessageId = channel.last_message_id {
-                return lastUnreadId < lastMessageId
-            }
-            return false
-        }()
-        
-        if (hasMentions) && hasUnread {
-            return .unreadWithMentions(
-                mentionsCount: formattedMentionCount(unread.mentions!.count)
-            )
-        } else if hasMentions {
-            return .mentions(formattedMentionCount(unread.mentions!.count))
-        } else if hasUnread {
-            return .unread
-        }
-
-        return nil
-        
-    }
-    
-    func getUnreadCountFor(server: Server) -> UnreadCount? {
-        if let serverNotificationValue = userSettingsStore.cache.notificationSettings.server[server.id] {
-            if serverNotificationValue == .muted && serverNotificationValue == .none {
-                return nil
-            }
-        }
-        
-        // FIXED: Use allEventChannels to check unreads for all channels, not just loaded ones
-        let serverChannelIds = server.channels
-        let channelUnreads = serverChannelIds.compactMap { channelId -> (Channel, UnreadCount?)? in
-            // First try from loaded channels, then from stored channels
-            if let channel = channels[channelId] {
-                return (channel, getUnreadCountFor(channel: channel))
-            } else if let channel = allEventChannels[channelId] {
-                // For unloaded channels, check unreads directly
-                let unread = unreads[channelId]
-                if let unread = unread {
-                    let unreadCount = getUnreadCountFromUnread(unread: unread, channel: channel)
-                    return (channel, unreadCount)
-                }
-                return (channel, nil)
-            }
-            return nil
-        }
-        
-        var mentionCount = 0
-        var hasUnread = false
-        
-        for (channel, unread) in channelUnreads {
-            let channelNotificationValue = userSettingsStore.cache.notificationSettings.channel[channel.id]
-            
-            if let unread = unread {
-                switch unread {
-                case .unread:
-                    if channelNotificationValue != NotificationState.none && channelNotificationValue != .muted {
-                        hasUnread = true
-                    }
-                    
-                case .mentions(let count):
-                    if channelNotificationValue != NotificationState.none && channelNotificationValue != .mention {
-                        mentionCount += (Int(count) ?? 0)
-                    }
-                case .unreadWithMentions(let count):
-                    if channelNotificationValue != NotificationState.none && channelNotificationValue != .mention {
-                        hasUnread = true
-                        mentionCount += (Int(count) ?? 0)
-                    }
-                    
-                }
-                
-            }
-        }
-        
-        if mentionCount > 0 && hasUnread {
-            return .unreadWithMentions(mentionsCount: formattedMentionCount(mentionCount))
-        }else if mentionCount > 0 {
-            return .mentions(formattedMentionCount(mentionCount))
-        } else if hasUnread {
-            return .unread
-        }
-        
-        return nil
-    }
-    
-    
-    func formattedMentionCount(_ input: Int) -> String {
-        if input > 10 {
-            return "+9"
-        } else {
-            return "\(input)"
-        }
-    }
-    
-    // Helper function to convert Unread object to UnreadCount for lazy loaded channels
-    func getUnreadCountFromUnread(unread: Unread, channel: Channel) -> UnreadCount? {
-        // Check channel notification settings
-        let channelNotificationValue = userSettingsStore.cache.notificationSettings.channel[channel.id]
-        
-        if channelNotificationValue == NotificationState.none || channelNotificationValue == .muted {
-            return nil
-        }
-        
-        // Check if channel has last message
-        guard let lastMessageId = channel.last_message_id else {
-            return nil
-        }
-        
-        // Check if there are unread messages
-        let hasUnreadMessages = unread.last_id != lastMessageId
-        
-        // Check for mentions
-        let mentionCount = unread.mentions?.count ?? 0
-        let hasMentions = mentionCount > 0
-        
-        if hasUnreadMessages && hasMentions {
-            return .unreadWithMentions(mentionsCount: formattedMentionCount(mentionCount))
-        } else if hasMentions {
-            return .mentions(formattedMentionCount(mentionCount))
-        } else if hasUnreadMessages {
-            return .unread
-        }
-        
-        return nil
-    }
-    
     func openUserSheet(withId id: String, server: String?) {
         if let user = users[id] {
             let member = server
@@ -4581,179 +2027,6 @@ public class ViewState: ObservableObject {
                 currentChannel = .home
             }
         }
-    }
-    
-    func selectServer(withId id: String) {
-        // Unload previous server's channels if switching servers
-        if case .server(let previousServerId) = currentSelection, previousServerId != id {
-            unloadServerChannels(serverId: previousServerId)
-        }
-        
-        currentSelection = .server(id)
-        
-        // LAZY LOADING: Load channels for this server
-        loadServerChannels(serverId: id)
-        
-        // CONDITIONAL: Only preload if automatic preloading is enabled
-        if enableAutomaticPreloading {
-            // PERFORMANCE: Start preloading messages for this server's channels
-            preloadMessagesForServer(serverId: id)
-            
-            // ENHANCED: Also trigger smart preloading for important channels
-            Task {
-                await preloadImportantChannels()
-            }
-            print("🚀 PRELOAD_ENABLED: Started automatic preloading for server \(id)")
-        } else {
-            print("📵 PRELOAD_DISABLED: Skipped automatic preloading for server \(id) channels")
-        }
-        
-        if let last = userSettingsStore.store.lastOpenChannels[id] {
-            currentChannel = .channel(last)
-        } else if let server = servers[id] {
-            if let firstChannel = server.channels.compactMap({
-                switch channels[$0] {
-                case .text_channel(let c):
-                    return c
-                default:
-                    return nil
-                }
-            }).first {
-                currentChannel = .channel(firstChannel.id)
-            } else {
-                currentChannel = .noChannel
-            }
-        }
-    }
-    
-    func selectChannel(inServer server: String, withId id: String) {
-        // Clear messages from previous channel before switching
-        if case .channel(let previousChannelId) = currentChannel, previousChannelId != id {
-            clearChannelMessages(channelId: previousChannelId)
-        }
-        
-        // CRITICAL FIX: Only clear target message ID if we're navigating to a DIFFERENT channel
-        // Don't clear it if we're navigating TO the target channel (from links/replies)
-        if let targetId = currentTargetMessageId {
-            // If target message is for current channel, keep it; otherwise clear it
-            if let targetMessage = messages[targetId], targetMessage.channel != id {
-                print("🎯 SELECT_CHANNEL: Clearing currentTargetMessageId - target is for different channel")
-                currentTargetMessageId = nil
-            } else if messages[targetId] == nil {
-                // Target message not loaded yet, assume it might be for this channel - keep it
-                print("🎯 SELECT_CHANNEL: Keeping currentTargetMessageId for channel \(id) - target message not loaded yet")
-            } else {
-                print("🎯 SELECT_CHANNEL: Keeping currentTargetMessageId for target channel \(id)")
-            }
-        }
-        
-        currentChannel = .channel(id)
-        userSettingsStore.store.lastOpenChannels[server] = id
-        
-        // CONDITIONAL: Only preload if automatic preloading is enabled
-        if enableAutomaticPreloading {
-            // AGGRESSIVE PRELOADING: Immediately preload this channel
-            Task {
-                await preloadSpecificChannel(channelId: id)
-            }
-            print("🚀 PRELOAD_ENABLED: Started automatic preloading for selected channel \(id)")
-        } else {
-            print("📵 PRELOAD_DISABLED: Skipped automatic preloading for selected channel \(id)")
-        }
-        
-        // CRITICAL FIX: Load users for visible messages when entering channel
-        loadUsersForVisibleMessages(channelId: id)
-    }
-    
-    func selectDms() {
-        DispatchQueue.main.async {
-            // Unload current server's channels when switching to DMs
-            if case .server(let serverId) = self.currentSelection {
-                self.unloadServerChannels(serverId: serverId)
-            }
-            
-            self.currentSelection = .dms
-            
-            if let last = self.userSettingsStore.store.lastOpenChannels["dms"] {
-                self.currentChannel = .channel(last)
-            } else {
-                // print("🏠 HOME_REDIRECT: Going to home because no last DM channel saved")
-                self.currentChannel = .home
-            }
-            
-            // FIX: Reinitialize DM list if it was cleared or not initialized
-            if !self.isDmListInitialized || self.dms.isEmpty {
-                self.reinitializeDmListFromCache()
-            }
-            
-            // CRITICAL FIX: Load users for visible messages when entering DM view
-            if case .channel(let channelId) = self.currentChannel {
-                self.loadUsersForVisibleMessages(channelId: channelId)
-            }
-        }
-    }
-    
-    func selectDiscover() {
-        DispatchQueue.main.async {
-            // Unload current server's channels when switching to Discover
-            if case .server(let serverId) = self.currentSelection {
-                self.unloadServerChannels(serverId: serverId)
-            }
-            
-            self.currentSelection = .discover
-            self.currentChannel = .home
-            
-            // Clear navigation path to go back to home/discover view
-            self.path.removeAll()
-        }
-    }
-    
-    @MainActor
-    func selectDm(withId id: String) {
-        // Clear messages from previous channel before switching
-        if case .channel(let previousChannelId) = self.currentChannel, previousChannelId != id {
-            self.clearChannelMessages(channelId: previousChannelId)
-        }
-        
-        // CRITICAL FIX: Only clear target message ID if we're navigating to a DIFFERENT channel
-        // Don't clear it if we're navigating TO the target channel (from links/replies)
-        if let targetId = currentTargetMessageId {
-            // If target message is for current channel, keep it; otherwise clear it
-            if let targetMessage = messages[targetId], targetMessage.channel != id {
-                print("🎯 SELECT_DM: Clearing currentTargetMessageId - target is for different channel")
-                currentTargetMessageId = nil
-            } else if messages[targetId] == nil {
-                // Target message not loaded yet, assume it might be for this channel - keep it
-                print("🎯 SELECT_DM: Keeping currentTargetMessageId for DM \(id) - target message not loaded yet")
-            } else {
-                print("🎯 SELECT_DM: Keeping currentTargetMessageId for target DM \(id)")
-            }
-        }
-        
-        self.currentChannel = .channel(id)
-        guard let channel = self.channels[id] else { return }
-        
-        switch channel {
-        case .dm_channel, .group_dm_channel:
-            self.userSettingsStore.store.lastOpenChannels["dms"] = id
-        default:
-            self.userSettingsStore.store.lastOpenChannels.removeValue(forKey: "dms")
-            
-        }
-        
-        // CONDITIONAL: Only preload if automatic preloading is enabled
-        if enableAutomaticPreloading {
-            // AGGRESSIVE PRELOADING: Immediately preload this DM
-            Task {
-                await preloadSpecificChannel(channelId: id)
-            }
-            print("🚀 PRELOAD_ENABLED: Started automatic preloading for selected DM \(id)")
-        } else {
-            print("📵 PRELOAD_DISABLED: Skipped automatic preloading for selected DM \(id)")
-        }
-        
-        // CRITICAL FIX: Load users for visible messages when entering DM
-        self.loadUsersForVisibleMessages(channelId: id)
     }
     
     func resolveAvatarUrl(user: Types.User, member: Member?, masquerade: Masquerade?) -> (url: URL, username: String, isAvatarSet : Bool) {
@@ -5010,218 +2283,12 @@ public class ViewState: ObservableObject {
     
     @Published var deviceNotificationToken: String? = nil
     @Published var pendingNotificationToken: String? = nil
-    
-    /// Retry uploading pending notification token
-    func retryUploadNotificationToken() async {
-        guard let token = pendingNotificationToken else { return }
-        
-        // print("🔄 RETRY_NOTIFICATION_TOKEN: Attempting to upload previously failed token...")
-        
-        let response = await http.uploadNotificationToken(token: token)
-        switch response {
-            case .success:
-                // print("✅ RETRY_NOTIFICATION_TOKEN: Successfully uploaded pending token")
-                pendingNotificationToken = nil // Clear pending token after success
-                UserDefaults.standard.removeObject(forKey: "pendingNotificationToken")
-            case .failure(let error):
-                print("❌ RETRY_NOTIFICATION_TOKEN: Failed again: \(error)")
-                // Keep the pending token for next retry
-        }
-    }
-    
-    /// Store notification token for later retry
-    func storePendingNotificationToken(_ token: String) {
-        pendingNotificationToken = token
-        UserDefaults.standard.set(token, forKey: "pendingNotificationToken")
-    }
-    
-    /// Load any pending notification token from storage
-    func loadPendingNotificationToken() {
-        pendingNotificationToken = UserDefaults.standard.string(forKey: "pendingNotificationToken")
-        if pendingNotificationToken != nil {
-            // print("📱 PENDING_TOKEN_FOUND: Found pending notification token to upload")
-        }
-    }
+
     
     // MARK: - Message Preloading System
     private let messagePreloadingQueue = DispatchQueue(label: "messagePreloading", qos: .utility)
-    private var preloadingTasks: [String: Task<Void, Never>] = [:]
-    private let maxPreloadedMessagesPerChannel = 20
-    
-    /// Preload recent messages for channels in a server to improve performance
-    private func preloadMessagesForServer(serverId: String) {
-        // Cancel any existing preloading task for this server
-        preloadingTasks[serverId]?.cancel()
-        
-        guard let server = servers[serverId] else { return }
-        
-        let task = Task { [weak self] in
-            guard let self = self else { return }
-            
-            // Get text channels that don't have messages loaded yet
-            let channelsToPreload = server.channels.compactMap { channelId -> String? in
-                guard let channel = self.channels[channelId] else { return nil }
-                
-                // Only preload text channels
-                switch channel {
-                case .text_channel(_):
-                    // Only preload if we don't have messages or have very few
-                    let existingMessageCount = self.channelMessages[channelId]?.count ?? 0
-                    return existingMessageCount < 5 ? channelId : nil
-                default:
-                    return nil
-                }
-            }
-            
-            // Limit concurrent preloading to avoid overwhelming the API
-            let maxConcurrentPreloads = 3
-            for channelBatch in channelsToPreload.chunked(into: maxConcurrentPreloads) {
-                await withTaskGroup(of: Void.self) { group in
-                    for channelId in channelBatch {
-                        group.addTask {
-                            await self.preloadChannelMessages(channelId: channelId, serverId: serverId)
-                        }
-                    }
-                }
-                
-                // Small delay between batches to be respectful to the API
-                try? await Task.sleep(nanoseconds: 500_000_000) // 0.5 seconds
-            }
-        }
-        
-        preloadingTasks[serverId] = task
-    }
-    
-    private func preloadChannelMessages(channelId: String, serverId: String) async {
-        // Check if task was cancelled
-        guard !Task.isCancelled else { return }
-        
-        do {
-            // SMART LIMIT: Use 10 for specific channel in specific server, 20 for others in preload
-            let messageLimit = (channelId == "01J7QTT66242A7Q26A2FH5TD48" && serverId == "01J544PT4T3WQBVBSDK3TBFZW7") ? 10 : maxPreloadedMessagesPerChannel
-            
-            let result = try await http.fetchHistory(
-                channel: channelId,
-                limit: messageLimit,
-                before: nil,
-                server: serverId,
-                messages: []
-            ).get()
-            
-            // Check if task was cancelled after API call
-            guard !Task.isCancelled else { return }
-            
-            await MainActor.run {
-                // Store users and members
-                for user in result.users {
-                    self.users[user.id] = user
-                }
-                
-                if let members = result.members {
-                    for member in members {
-                        self.members[member.id.server, default: [:]][member.id.user] = member
-                    }
-                }
-                
-                // Store messages
-                var messageIds: [String] = []
-                for message in result.messages {
-                    self.messages[message.id] = message
-                    messageIds.append(message.id)
-                }
-                
-                // Store message IDs in channel (sorted by creation time)
-                let sortedIds = messageIds.sorted { id1, id2 in
-                    let date1 = createdAt(id: id1)
-                    let date2 = createdAt(id: id2)
-                    return date1 < date2
-                }
-                
-                self.channelMessages[channelId] = sortedIds
-                
-                // print("📥 PRELOAD: Cached \(result.messages.count) messages for channel \(channelId)")
-            }
-        } catch {
-            // Silently handle errors - preloading is a performance optimization, not critical
-            // print("⚠️ PRELOAD: Failed to preload messages for channel \(channelId): \(error)")
-        }
-    }
-    
-    // MARK: - App Badge Management
-    
-    /// Calculates the total unread count across all channels and updates the app badge
-    func updateAppBadgeCount() {
-        guard let application = ViewState.application else { return }
-        
-        var totalUnreadCount = 0
-        var totalMentionCount = 0
-        
-        // Iterate through all unreads
-        for (channelId, unread) in unreads {
-            // Get channel info
-            let channel = channels[channelId] ?? allEventChannels[channelId]
-            
-            // Skip if channel doesn't exist
-            guard let channel = channel else {
-                continue
-            }
-            
-            // Check if channel is muted
-            let channelNotificationState = userSettingsStore.cache.notificationSettings.channel[channelId]
-            let isChannelMuted = channelNotificationState == .muted || channelNotificationState == .none
-            
-            // Check if server is muted (only for server channels, not DMs or group DMs)
-            var isServerMuted = false
-            if let serverId = channel.server {
-                let serverNotificationState = userSettingsStore.cache.notificationSettings.server[serverId]
-                isServerMuted = serverNotificationState == .muted || serverNotificationState == .none
-            }
-            // For DMs and group DMs, server is nil, so isServerMuted stays false
-            
-            // Skip if channel or server is muted
-            if isChannelMuted || isServerMuted {
-                continue
-            }
-            
-            // Count unread channels (including group DMs)
-            if let lastUnreadId = unread.last_id, let lastMessageId = channel.last_message_id {
-                if lastUnreadId < lastMessageId {
-                    totalUnreadCount += 1
-                    
-                    // Debug log for group DMs
-                    if case .group_dm_channel(let groupDM) = channel {
-                        print("🔔 Badge: Counting group DM '\(groupDM.name)' as unread")
-                    }
-                }
-            }
-        }
-        
-        // Total badge count is only unread channels (not mentions)
-        let finalBadgeCount = totalUnreadCount
-        
-        // Update app badge count
-        DispatchQueue.main.async {
-            let currentBadge = application.applicationIconBadgeNumber
-            application.applicationIconBadgeNumber = finalBadgeCount
-            print("🔔 Badge: \(currentBadge) -> \(finalBadgeCount) (unreads: \(totalUnreadCount))")
-        }
-    }
-    
-    /// Clears the app badge count
-    func clearAppBadge() {
-        guard let application = ViewState.application else { return }
-        
-        DispatchQueue.main.async {
-            application.applicationIconBadgeNumber = 0
-            print("🔔 Cleared app badge count")
-        }
-    }
-    
-    /// Manually refreshes the app badge count - useful for debugging or when the count seems incorrect
-    func refreshAppBadge() {
-        print("🔔 Manually refreshing app badge count...")
-        updateAppBadgeCount()
-    }
+    internal var preloadingTasks: [String: Task<Void, Never>] = [:]
+    internal let maxPreloadedMessagesPerChannel = 20
     
     /// Debug badge count and print detailed analysis to console
     func debugBadgeCount() {
@@ -5284,230 +2351,8 @@ public class ViewState: ObservableObject {
         print("  - Total channels stored: \(allEventChannels.count)")
         print("🔍 === END DEBUG ===\n")
     }
-    
-    /// Clean up stale unread entries for channels that no longer exist
-    func cleanupStaleUnreads() {
-        print("🧹 Cleaning up stale unreads...")
-        var removedCount = 0
-        var staleChannels: [String] = []
-        
-        for channelId in unreads.keys {
-            // Check if channel exists in our channels dictionary or allEventChannels
-            if channels[channelId] == nil && allEventChannels[channelId] == nil {
-                staleChannels.append(channelId)
-                removedCount += 1
-            }
-        }
-        
-        // Remove stale entries
-        for channelId in staleChannels {
-            unreads.removeValue(forKey: channelId)
-            print("  ❌ Removed stale unread for channel: \(channelId)")
-        }
-        
-        print("🧹 Cleanup complete. Removed \(removedCount) stale entries.")
-        
-        // Update badge count after cleanup
-        updateAppBadgeCount()
-    }
-    
-    /// Force mark all channels as read and clear the app badge
-    func forceMarkAllAsRead() {
-        print("📖 Force marking all channels as read...")
-        let channelCount = unreads.count
-        
-        // Clear all unreads
-        unreads.removeAll()
-        
-        // Clear the app badge
-        clearAppBadge()
-        
-        print("📖 Marked \(channelCount) channels as read and cleared badge")
-    }
-    
-    /// Show detailed unread message counts for each channel
-    func showUnreadCounts() {
-        print("\n📊 === UNREAD MESSAGE COUNTS ===")
-        
-        var totalUnreadMessages = 0
-        var totalMentions = 0
-        var channelsWithUnread: [(name: String, id: String, unreadCount: Int, mentionCount: Int)] = []
-        
-        for (channelId, unread) in unreads {
-            let channel = channels[channelId] ?? allEventChannels[channelId]
-            let channelName = channel?.name ?? "Unknown Channel"
-            
-            // Skip if channel doesn't exist
-            guard let channel = channel else {
-                print("❌ Channel \(channelId) not found - skipping")
-                continue
-            }
-            
-            // Check notification settings
-            let isChannelMuted = userSettingsStore.cache.notificationSettings.channel[channelId] == .muted || 
-                                userSettingsStore.cache.notificationSettings.channel[channelId] == .none
-            let serverIdForChannel = channel.server
-            let isServerMuted = serverIdForChannel != nil ? 
-                (userSettingsStore.cache.notificationSettings.server[serverIdForChannel!] == .muted || 
-                 userSettingsStore.cache.notificationSettings.server[serverIdForChannel!] == .none) : false
-            
-            // Calculate unread count
-            var unreadCount = 0
-            if let lastUnreadId = unread.last_id, let lastMessageId = channel.last_message_id {
-                if lastUnreadId < lastMessageId {
-                    // We can't get exact count without fetching messages, but we know there are unread messages
-                    unreadCount = -1 // -1 means "has unread but count unknown"
-                }
-            }
-            
-            let mentionCount = unread.mentions?.count ?? 0
-            
-            if unreadCount != 0 || mentionCount > 0 {
-                let mutedIndicator = (isChannelMuted || isServerMuted) ? " 🔇" : ""
-                channelsWithUnread.append((
-                    name: channelName + mutedIndicator,
-                    id: channelId,
-                    unreadCount: unreadCount,
-                    mentionCount: mentionCount
-                ))
-                
-                if !(isChannelMuted || isServerMuted) {
-                    if unreadCount == -1 {
-                        totalUnreadMessages += 1 // Count as at least 1
-                    } else if unreadCount > 0 {
-                        totalUnreadMessages += unreadCount
-                    }
-                    totalMentions += mentionCount
-                }
-            }
-        }
-        
-        // Sort by mention count first, then by name
-        channelsWithUnread.sort { 
-            if $0.mentionCount != $1.mentionCount {
-                return $0.mentionCount > $1.mentionCount
-            }
-            return $0.name < $1.name
-        }
-        
-        // Print results
-        if channelsWithUnread.isEmpty {
-            print("✅ No channels with unread messages!")
-        } else {
-            print("\n📌 Channels with unread messages:")
-            for channel in channelsWithUnread {
-                let unreadText = channel.unreadCount == -1 ? "Has unread" : "\(channel.unreadCount) unread"
-                let mentionText = channel.mentionCount > 0 ? ", \(channel.mentionCount) mention(s)" : ""
-                print("  • \(channel.name): \(unreadText)\(mentionText)")
-            }
-        }
-        
-        print("\n📊 Summary:")
-        print("  - Total channels with unread: \(channelsWithUnread.count)")
-        print("  - Total unread channels (unmuted): \(totalUnreadMessages)")
-        print("  - Total mentions (unmuted): \(totalMentions)")
-        print("  - Current badge count: \(ViewState.application?.applicationIconBadgeNumber ?? 0)")
-        print("📊 === END UNREAD COUNTS ===\n")
-    }
-    
-    /// Get unread counts as a formatted string for UI display
-    func getUnreadCountsString() -> String {
-        var result = "📊 UNREAD MESSAGE COUNTS\n\n"
-        
-        var channelsWithUnread: [(name: String, id: String, unreadCount: Int, mentionCount: Int, isMuted: Bool)] = []
-        
-        for (channelId, unread) in unreads {
-            let channel = channels[channelId] ?? allEventChannels[channelId]
-            let channelName = channel?.name ?? "Unknown Channel"
-            
-            // Skip if channel doesn't exist
-            guard let channel = channel else {
-                continue
-            }
-            
-            // Check notification settings
-            let isChannelMuted = userSettingsStore.cache.notificationSettings.channel[channelId] == .muted || 
-                                userSettingsStore.cache.notificationSettings.channel[channelId] == .none
-            let serverIdForChannel = channel.server
-            let isServerMuted = serverIdForChannel != nil ? 
-                (userSettingsStore.cache.notificationSettings.server[serverIdForChannel!] == .muted || 
-                 userSettingsStore.cache.notificationSettings.server[serverIdForChannel!] == .none) : false
-            
-            let isMuted = isChannelMuted || isServerMuted
-            
-            // Calculate unread count
-            var unreadCount = 0
-            if let lastUnreadId = unread.last_id, let lastMessageId = channel.last_message_id {
-                if lastUnreadId < lastMessageId {
-                    unreadCount = -1 // -1 means "has unread but count unknown"
-                }
-            }
-            
-            let mentionCount = unread.mentions?.count ?? 0
-            
-            if unreadCount != 0 || mentionCount > 0 {
-                channelsWithUnread.append((
-                    name: channelName,
-                    id: channelId,
-                    unreadCount: unreadCount,
-                    mentionCount: mentionCount,
-                    isMuted: isMuted
-                ))
-            }
-        }
-        
-        // Sort by muted status first, then mention count, then name
-        channelsWithUnread.sort { 
-            if $0.isMuted != $1.isMuted {
-                return !$0.isMuted // Unmuted first
-            }
-            if $0.mentionCount != $1.mentionCount {
-                return $0.mentionCount > $1.mentionCount
-            }
-            return $0.name < $1.name
-        }
-        
-        if channelsWithUnread.isEmpty {
-            result += "✅ No channels with unread messages!"
-        } else {
-            var unmutedCount = 0
-            var mutedCount = 0
-            
-            result += "📌 Unmuted channels:\n"
-            for channel in channelsWithUnread where !channel.isMuted {
-                let unreadText = channel.unreadCount == -1 ? "Has unread" : "\(channel.unreadCount) unread"
-                let mentionText = channel.mentionCount > 0 ? ", \(channel.mentionCount) mention(s)" : ""
-                result += "• \(channel.name): \(unreadText)\(mentionText)\n"
-                unmutedCount += 1
-            }
-            
-            if unmutedCount == 0 {
-                result += "None\n"
-            }
-            
-            result += "\n🔇 Muted channels:\n"
-            for channel in channelsWithUnread where channel.isMuted {
-                let unreadText = channel.unreadCount == -1 ? "Has unread" : "\(channel.unreadCount) unread"
-                let mentionText = channel.mentionCount > 0 ? ", \(channel.mentionCount) mention(s)" : ""
-                result += "• \(channel.name): \(unreadText)\(mentionText)\n"
-                mutedCount += 1
-            }
-            
-            if mutedCount == 0 {
-                result += "None\n"
-            }
-            
-            result += "\n📊 Summary:\n"
-            result += "• Total channels with unread: \(channelsWithUnread.count)\n"
-            result += "• Unmuted channels: \(unmutedCount)\n"
-            result += "• Muted channels: \(mutedCount)\n"
-            result += "• Current badge: \(ViewState.application?.applicationIconBadgeNumber ?? 0)"
-        }
-        
-        return result
-    }
     // MARK: - Message queue processing state
-    private var isProcessingQueue: [String: Bool] = [:] // Prevent concurrent processing per channel
+    internal var isProcessingQueue: [String: Bool] = [:] // Prevent concurrent processing per channel
     
     // Check the internet status
     func setupInternetObservation() {
@@ -5520,95 +2365,6 @@ public class ViewState: ObservableObject {
             }
             .store(in: &cancellables)
         }
-    
-    // Function to send the queuing of messages
-    func trySendingQueuedMessages() {
-        print("👍🏻 Entered trySendingQueuedMessages")
-        guard InternetMonitor.shared.isConnected else { 
-            print("❌ Not connected, aborting queue send")
-            return 
-        }
-
-        print("📌 queuedMessages count:", queuedMessages.count)
-        
-        // Get a snapshot of channels to process (avoid concurrent modification)
-        let channelsToProcess = Array(queuedMessages.keys)
-        
-        // Process each channel sequentially with concurrency guard
-        for channelId in channelsToProcess {
-            // Skip if already processing this channel
-            if isProcessingQueue[channelId] == true {
-                print("⏭️ Skipping channel \(channelId) - already processing")
-                continue
-            }
-            
-            // Create a task for this channel to maintain message order
-            Task {
-                // Mark channel as being processed
-                await MainActor.run {
-                    self.isProcessingQueue[channelId] = true
-                }
-                
-                defer {
-                    // Always clear the processing flag when done
-                    Task { @MainActor in
-                        self.isProcessingQueue[channelId] = false
-                    }
-                }
-                
-                var sentCount = 0
-                
-                // Keep sending from the front of the queue until it's empty or send fails
-                while await MainActor.run(body: { self.queuedMessages[channelId]?.isEmpty == false }) {
-                    // Safely get and remove the first message atomically
-                    let msg = await MainActor.run { () -> QueuedMessage? in
-                        guard let first = self.queuedMessages[channelId]?.first else {
-                            return nil
-                        }
-                        // Remove it immediately to prevent duplicate sends
-                        self.queuedMessages[channelId]?.removeFirst()
-                        return first
-                    }
-                    
-                    guard let msg = msg else {
-                        break
-                    }
-                    
-                    sentCount += 1
-                    print("📌 Sending queued message \(sentCount) for channel \(channelId) - nonce: \(msg.nonce)")
-                    
-                    do {
-                        print("👍🏻 Sending message: \(msg.content)")
-                        let _ = try await http.sendMessage(
-                            channel: channelId,
-                            replies: msg.replies,
-                            content: msg.content,
-                            attachments: [],
-                            nonce: msg.nonce
-                        ).get()
-                        print("📤 Sent queued message \(sentCount) successfully - nonce: \(msg.nonce)")
-                    } catch {
-                        print("❌ Failed to send queued message - nonce: \(msg.nonce), error: \(error)")
-                        // Re-add the message back to the front of the queue on failure
-                        await MainActor.run {
-                            self.queuedMessages[channelId]?.insert(msg, at: 0)
-                        }
-                        // Stop trying to send remaining messages in this channel if one fails
-                        break
-                    }
-                }
-                
-                // Clean up empty queue entries
-                await MainActor.run {
-                    if self.queuedMessages[channelId]?.isEmpty == true {
-                        self.queuedMessages.removeValue(forKey: channelId)
-                    }
-                    print("👍🏻 Finished processing channel \(channelId) - sent \(sentCount) messages")
-                }
-            }
-        }
-        print("👍🏻 Exiting trySendingQueueMessage")
-    }
 }
 
 // This is used for notifications @mention issue fetches the corresponding @mention user from users.json. Testing is pending.
@@ -5631,397 +2387,6 @@ public class ViewState: ObservableObject {
 //    }
 //}
 
-
-extension ViewState {
-    
-    // Defining the path of the cache and type of the cache
-    static func serversCacheURL() -> URL? {
-        let fileManager = FileManager.default
-        guard let appSupport = fileManager.urls(for: .applicationSupportDirectory, in: .userDomainMask).first else { return nil }
-        let dir = appSupport.appendingPathComponent(Bundle.main.bundleIdentifier ?? "ZekoChat")
-        do {
-            try fileManager.createDirectory(at: dir, withIntermediateDirectories: true)
-        } catch {
-            print("❌ Failed to create Application Support directory:", error)
-            return nil
-        }
-        return dir.appendingPathComponent("servers_cache.json")
-    }
-    
-    // Load the cache when app boots up
-    static func loadServersCacheSync() -> OrderedDictionary <String, Server> {
-        guard let url = serversCacheURL(), FileManager.default.fileExists(atPath: url.path) else {
-            return [:]
-        }
-        do {
-            let data  = try Data(contentsOf: url)
-            let decoder = JSONDecoder()
-            return try decoder.decode(OrderedDictionary<String, Server>.self, from: data)
-        } catch {
-            print("❌ Failed to load servers cache:", error)
-            return [:]
-        }
-    }
-    
-    func saveServersCacheAsync() {
-        let serversSnapshot = self.servers
-        Task.detached(priority: .background) {
-            do {
-                let encoder = JSONEncoder()
-                let data = try encoder.encode(serversSnapshot)
-                if let url = await ViewState.serversCacheURL() {
-                    try data.write(to: url, options: .atomic)
-                    print("✅ Saved servers cache to \(url.path)")
-                }
-            } catch {
-                print("❌ Failed to write servers cache:", error)
-            }
-        }
-    }
-    
-    
-    func applyServerOrdering() {
-        let ordering = self.userSettingsStore.cache.orderSettings.servers
-        let allServers = Array(self.servers.values)
-
-        let serverDict = Dictionary(uniqueKeysWithValues: allServers.map { ($0.id, $0) })
-        let orderedServers = ordering.compactMap { serverDict[$0] }
-        let remainingServers = allServers.filter { !ordering.contains($0.id) }
-
-        let finalServers = orderedServers + remainingServers
-
-        // Update `servers` preserving key-value order
-        var newServers = OrderedDictionary<String, Server>()
-        for server in finalServers {
-            newServers[server.id] = server
-        }
-        self.servers = newServers
-    }
-}
-
-
-extension ViewState {
-    /// Sets the `active` property of a DMChannel with the given ID to `false`.
-    /// - Parameter id: The ID of the channel to update.
-    func deactivateDMChannel(with id: String) {
-        Task { @MainActor in
-            if let index = dms.firstIndex(where: {
-                if case let .dm_channel(dmChannel) = $0 {
-                    return dmChannel.id == id && dmChannel.active
-                }
-                return false
-            }) {
-                if case var .dm_channel(dmChannel) = dms[index] {
-                    // Update the active state to false
-                    dmChannel.active = false
-                    dms[index] = .dm_channel(dmChannel) // Update the channel in the list
-                }
-            }
-        }
-    }
-    
-    /// Closes the DM group by calling an API and deactivating the DM channel in the `dms` list.
-    /// - Parameter channelId: The ID of the channel to close.
-    func closeDMGroup(channelId: String) async {
-        do {
-            // Call the API to close the DM group
-            let _ = try await self.http.closeDMGroup(channelId: channelId).get()
-            
-            // Deactivate the DM channel in the list
-            await MainActor.run {
-                self.deactivateDMChannel(with: channelId)
-            }
-        } catch let error {
-            print("Error closing DM group: \(error)")
-        }
-    }
-    
-    
-    func isCurrentUserOwner(of serverID: String) -> Bool {
-        guard let currentUser = currentUser else {
-            return false
-        }
-        
-        guard let server = servers[serverID] else {
-            return false
-        }
-        
-        return server.owner == currentUser.id
-    }
-    
-    
-    func removeServer(with serverID: String) {
-        Task { @MainActor in
-            servers.removeValue(forKey: serverID)
-            selectDms()
-        }
-    }
-    
-    @MainActor
-    func removeChannel(with channelID : String, initPath: Bool = true){
-        if(initPath){
-            self.path = .init()
-            selectDms()
-        }
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
-            //TODO
-            self.channels.removeValue(forKey: channelID)
-            self.dms = self.dms.filter { $0.id != channelID }
-        }
-        
-    }
-    
-    // MARK: - Ready Event Processing Functions
-    
-    // Structure to hold only the data we need
-    private struct ReadyEventData {
-        let channels: [Channel]
-        let servers: [Server]
-        let users: [Types.User]
-        let members: [Member]
-        let emojis: [Emoji]
-    }
-    
-    // Extract only needed data from the large event
-    private func extractNeededDataFromReadyEvent(_ event: ReadyEvent) -> ReadyEventData {
-        // print("🚀 VIEWSTATE: Extracting needed data from ready event")
-        
-        // Process all servers (removed limitation)
-        let ordering = self.userSettingsStore.cache.orderSettings.servers
-        let serverDict = Dictionary(uniqueKeysWithValues: event.servers.map { ($0.id, $0) })
-        let orderedServers: [Server] = ordering.compactMap { serverDict[$0] }
-        let remainingServers = event.servers.filter { !ordering.contains($0.id) }
-        let allServers = orderedServers + remainingServers
-        
-        // print("   - Processing all \(event.servers.count) servers (no limitation)")
-        
-        // Get server IDs for channel filtering
-        let serverIds = Set(allServers.map { $0.id })
-        
-        // LAZY LOADING: Store ALL channels but only load DMs immediately
-        allEventChannels.removeAll() // Clear existing stored channels
-        
-        var neededChannels: [Channel] = [] // Only DMs will be loaded immediately
-        var dmCount = 0
-        var storedServerChannels = 0
-        
-        for channel in event.channels {
-            // Store ALL channels for lazy loading
-            allEventChannels[channel.id] = channel
-            
-            var shouldLoadNow = false
-            
-            switch channel {
-            case .dm_channel(let dm):
-                // ALWAYS load DMs immediately (both active and inactive)
-                shouldLoadNow = true
-                dmCount += 1
-                // print("🚀 VIEWSTATE: Loading DM channel \(channel.id) immediately (active: \(dm.active))")
-            case .group_dm_channel:
-                // ALWAYS load Group DMs immediately
-                shouldLoadNow = true
-                dmCount += 1
-                // print("🚀 VIEWSTATE: Loading Group DM channel \(channel.id) immediately")
-            case .text_channel(let textChannel):
-                // Store server channels but don't load them yet
-                if serverIds.contains(textChannel.server) {
-                    storedServerChannels += 1
-                    // print("🔄 LAZY_CHANNEL: Stored text channel \(channel.id) for server \(textChannel.server)")
-                }
-            case .voice_channel(let voiceChannel):
-                // Store voice channels but don't load them yet
-                if serverIds.contains(voiceChannel.server) {
-                    storedServerChannels += 1
-                    // print("🔄 LAZY_CHANNEL: Stored voice channel \(channel.id) for server \(voiceChannel.server)")
-                }
-            default:
-                break
-            }
-            
-            if shouldLoadNow {
-                neededChannels.append(channel)
-            }
-        }
-        
-        // print("   - Stored \(allEventChannels.count) total channels for lazy loading")
-        // print("   - Loading immediately: \(neededChannels.count) channels (DMs: \(dmCount))")
-        // print("   - Stored for lazy loading: \(storedServerChannels) server channels")
-        
-        // Return only the data we need
-        return ReadyEventData(
-            channels: neededChannels,
-            servers: allServers,
-            users: event.users, // Keep all users for now
-            members: event.members,
-            emojis: event.emojis
-        )
-    }
-    
-    // Process the extracted data
-    private func processReadyData(_ data: ReadyEventData) async {
-        let processReadySpan = launchTransaction?.startChild(operation: "processReady")
-        
-        // Process channels
-        processChannelsFromData(data.channels)
-        
-        // Process servers  
-        processServersFromData(data.servers)
-        
-        // Process users
-        processUsers(data.users)
-        
-        // EMERGENCY: If still too many users, force immediate cleanup
-        if users.count > maxUsersInMemory {
-            // print("🚨 EMERGENCY: Still have \(users.count) users after processing, forcing cleanup!")
-            let currentUserId = currentUser?.id
-            let currentUserObject = currentUser
-            
-            users.removeAll()
-            
-            if let currentUserId = currentUserId, let currentUserObject = currentUserObject {
-                users[currentUserId] = currentUserObject
-                currentUser = currentUserObject
-            }
-            
-            // print("🚨 EMERGENCY CLEANUP: Reduced users to \(users.count)")
-        }
-        
-        // Process members
-        processMembers(data.members)
-        
-        // Process DMs
-        processDMs(channels: Array(channels.values))
-        
-        // Process emojis
-        for emoji in data.emojis {
-            self.emojis[emoji.id] = emoji
-        }
-        
-        // MEMORY FIX: Don't fetch unreads here - it might trigger another ready event
-        // We'll fetch them separately after full initialization
-        // print("🚀 VIEWSTATE: Skipping unreads fetch during ready processing to prevent memory spike")
-        
-        // Update state
-        state = .connected
-        wsCurrentState = .connected
-        ws?.currentState = .connected
-        ws?.retryCount = 0
-        
-        await verifyStateIntegrity()
-        
-        processReadySpan?.finish()
-        launchTransaction?.finish()
-        
-        // Check for stale messages
-        for channel in channels.values {
-            if let last_message_id = channel.last_message_id,
-               let last_cached_message = channelMessages[channel.id]?.last,
-               last_message_id != last_cached_message
-            {
-                channelMessages[channel.id] = []
-            }
-        }
-        
-        // print("🚀 VIEWSTATE: Ready event processing completed")
-        // print("   - Final channels: \(channels.count)")
-        // print("   - Final users: \(users.count)")
-        // print("   - Final servers: \(servers.count)")
-        
-        // Retry any pending notification token upload
-        if pendingNotificationToken != nil {
-            // print("🔄 READY_EVENT: Found pending notification token, attempting retry...")
-            Task {
-                await retryUploadNotificationToken()
-            }
-        }
-        
-        // MEMORY FIX: Fetch unreads separately to prevent memory spike
-        Task {
-            // print("🚀 VIEWSTATE: Starting unreads fetch after ready completion")
-            if let remoteUnreads = try? await http.fetchUnreads().get() {
-                await MainActor.run {
-                    for unread in remoteUnreads {
-                        unreads[unread.id.channel] = unread
-                    }
-                    // print("🚀 VIEWSTATE: Unreads loaded: \(remoteUnreads.count)")
-                    
-                    // Update app badge count after loading unreads from server
-                    updateAppBadgeCount()
-                    print("🔔 Updated badge count after loading \(remoteUnreads.count) unreads from server (mentions not counted)")
-                }
-            }
-        }
-    }
-    
-    private func processChannelsFromData(_ eventChannels: [Channel]) {
-        // print("🚀 VIEWSTATE: Processing \(eventChannels.count) channels from WebSocket")
-        // print("🚀 VIEWSTATE: Existing channels count: \(channels.count)")
-        
-        // CRITICAL FIX: Don't clear existing channels - merge instead
-        // Only clear channelMessages as they should be fresh from server
-        channelMessages.removeAll()
-        
-        var messageArrayCount = 0
-        var dmChannels = 0
-        var groupDmChannels = 0
-        var textChannels = 0
-        
-        // Process all channels (already filtered)
-        for channel in eventChannels {
-            channels[channel.id] = channel
-            
-            // Create message array for messageable channels
-            switch channel {
-            case .dm_channel(let dm):
-                // Create message array for ALL DMs (active and inactive)
-                channelMessages[channel.id] = []
-                messageArrayCount += 1
-                dmChannels += 1
-                // print("🚀 VIEWSTATE: Added DM channel \(channel.id) (active: \(dm.active))")
-            case .group_dm_channel:
-                channelMessages[channel.id] = []
-                messageArrayCount += 1
-                groupDmChannels += 1
-                // print("🚀 VIEWSTATE: Added Group DM channel \(channel.id)")
-            case .text_channel:
-                channelMessages[channel.id] = []
-                messageArrayCount += 1
-                textChannels += 1
-            default:
-                break
-            }
-        }
-        
-        // print("🚀 VIEWSTATE: Stored \(channels.count) channels, created \(messageArrayCount) message arrays")
-        // print("🚀 VIEWSTATE: Channel breakdown - DMs: \(dmChannels), Group DMs: \(groupDmChannels), Text: \(textChannels)")
-    }
-    
-    private func processServersFromData(_ servers: [Server]) {
-        // print("🚀 VIEWSTATE: Processing \(servers.count) servers from WebSocket")
-        // print("🚀 VIEWSTATE: Existing servers count: \(self.servers.count)")
-        
-        // CRITICAL FIX: Don't clear existing servers - merge instead
-        // Keep existing servers and update/add new ones from WebSocket
-        for server in servers {
-            self.servers[server.id] = server
-            if members[server.id] == nil {
-                members[server.id] = [:]
-            }
-        }
-        
-        // print("🚀 VIEWSTATE: Final servers count after merge: \(self.servers.count)")
-        // Apply ordering before saving cache to ensure correct order is persisted
-        self.applyServerOrdering()
-        self.saveServersCacheAsync()
-    }
-}
-
-
-enum UnreadCount : Equatable {
-    case unread
-    case mentions(String)
-    case unreadWithMentions(mentionsCount: String)
-}
 
 extension Dictionary {
     mutating func setDefault(key: Key, default def: Value) -> Value {
